### 2020-06-08 #2301 @kinnala
> [ point_right Check article proof page_facing_up point_left ](https://github.com/openjournals/joss-papers/blob/joss.02301/joss.02301/10.21105.joss.02301.pdf)

The proof looks correct. My coauthor @gdmcbain can also confirm.
