### 2020-06-09 #2301 @gdmcbain
> The proof looks correct. My coauthor @gdmcbain can also confirm.

LGTM too, thank you.
