### 2020-06-10 8813683
Update tests/test_examples: fix flake8 errors
```
tests/test_examples.py:51:9: F401 'docs.examples.ex08' imported but unused
tests/test_examples.py:117:9: F401 'docs.examples.ex17 as ex' imported but unused
tests/test_examples.py:123:9: F401 'docs.examples.ex18 as ex' imported but unused
tests/test_examples.py:129:9: F401 'docs.examples.ex19 as ex' imported but unused
tests/test_examples.py:159:4: E111 indentation is not a multiple of four
tests/test_examples.py:160:7: E111 indentation is not a multiple of four
tests/test_examples.py:161:7: E111 indentation is not a multiple of four
tests/test
[…]
