### 2020-06-11 76efb92
Update tests/test_elements.py: fix flake8 errors
```
tests/test_elements.py:6:1: F403 'from skfem import *' used; unable to detect undefined names
tests/test_elements.py:13:9: F405 'ElementLineP1' may be undefined, or defined from star imports: skfem
tests/test_elements.py:14:9: F405 'ElementLineP2' may be undefined, or defined from star imports: skfem
tests/test_elements.py:15:9: F405 'ElementLinePp' may be undefined, or defined from star imports: skfem
tests/test_elements.py:16:9: F405 'ElementLinePp' may be undefined, or defined from star imports: skfem
tests/test_elements.py:17:9: F405 'El
[…]
