### 2020-06-16 c019f14
Update tests/test_basis.py: fix flake8 errors
