### 2020-06-17 01ebaae
Create tests/test_element_quad_bfs.py
Add missing test.
