### 2020-06-19 #2301 @meg-simula
@yangbai90, @AnjaliSandip,  @capitalaslash, @thelfer Would you like to review this submission for JOSS? Feel free to take a look at our reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html .
