### 2020-06-19 #2301 @meg-simula
Thanks for the prompt response @thelfer. JOSS is currently in a "reduced service mode". For reviews, this means in particular that (cf. https://blog.joss.theoj.org/2020/05/reopening-joss):

> Revised (extended) review time: Before the suspension, we encouraged reviewers to complete their reviews in two weeks. In this reduced service mode, we’re asking that reviewers try to complete their reviews in six weeks. Reviewers are of course welcome to complete their reviews faster than this and given the iterative nature of JOSS reviews we still encourage everyone to start their review as soon as possible.
