### 2020-06-19 #2369 @whedon
[REVIEW]: scikit-fem: A Python package for finite element assembly
**Submitting author:** @kinnala (<a href="http://orcid.org/0000-0003-1611-5032">Tom Gustafsson</a>)
**Repository:** <a href="https://github.com/kinnala/scikit-fem" target ="_blank">https://github.com/kinnala/scikit-fem</a>
**Version:** v1.1.0
