### 2020-06-21 #2369 @AnjaliSandip
Hello,

This is my first time reviewing for JOSS. In regards to the checklist, do I select the boxes (if the paper meets that criterion) in the checklist under my name listed in this thread?

Thanks for any information you can provide.

Anjali
