### 2020-06-21 #2369 @danielskatz
Yes, that's correct.  If you have any problem doing so, be sure you have accepted the invite at this URL: https://github.com/openjournals/joss-reviews/invitations
