### 2020-06-21 #2369 @AnjaliSandip
Thanks for letting me know!

On Sun, Jun 21, 2020 at 11:56 AM Daniel S. Katz <notifications@github.com>
wrote:

> Yes, that's correct. If you have any problem doing so, be sure you have
> accepted the invite at this URL:
> https://github.com/openjournals/joss-reviews/invitations
>
