### 2020-06-29 704e7c3
first attempt at ElementLineMini
