### 2020-06-29 b97dbbf
optional ElementLineMini #413
