### 2020-06-29 19edcac
allow selectional of any element_line for velocity
