### 2020-07-03 708c70c
generalize line elements to other than P1
