### 2020-07-03 3077b94
remove old-style bilinear_form
