### 2020-07-09 b3f2617
Enable the merging of Dofs objects
