### 2020-07-09 91b42e5
Add test and changelog entry
