+++ tests/test_basis.py
@@ -5,7 +5,7 @@ from numpy.testing import assert_allclose
 
 from skfem import BilinearForm, asm, solve, condense
 from skfem.mesh import MeshTri, MeshTet, MeshHex
-from skfem.assembly import InteriorBasis
+from skfem.assembly import InteriorBasis, Dofs
 from skfem.element import (ElementVectorH1, ElementTriP2, ElementTriP1,
                            ElementTetP2, ElementHexS2)
 
@@ -89,3 +89,16 @@ class TestFacetExpansion(TestCase):
 class TestFacetExpansionHexS2(TestFacetExpansion):
     mesh_type = MeshHex
     elem_type = ElementHexS2
+
+
+class TestDofsMerge(TestCase):
+
+    def runTest(self):
+        m = MeshTri()
+        basis = InteriorBasis(m, ElementTriP2())
+        D1 = basis.get_dofs(lambda x: x[0] == 0)
+        D2 = basis.get_dofs(lambda x: x[0] == 1)
+        D3 = basis.get_dofs(lambda x: x[1] == 1)
+        D4 = basis.get_dofs(lambda x: x[1] == 0)
+        assert_allclose(D1 | D2 | D3 | D4,
+                        basis.get_dofs())
