### 2020-07-11 d4f54d3
Fix flake8 complaints
