+++ docs/examples/ex35.py
@@ -1,12 +1,17 @@
-"""Computation of the characteristic impedance and velocity factor of RG316 coaxial cable.
+"""Computation of the characteristic impedance and velocity factor of RG316
+coaxial cable.
 
 .. note::
-   This example requires the external package `pygmsh <https://pypi.org/project/pygmsh/>`_.
+   This example requires the external package
+   `pygmsh <https://pypi.org/project/pygmsh/>`_.
 
-This example solves the series inductance (per meter) and parallel capacitance (per meter)
-of RG316 coaxial cable. These values are then used to compute the characteristic impedance and velocity factor of the cable.
+This example solves the series inductance (per meter) and parallel capacitance
+(per meter)
+of RG316 coaxial cable. These values are then used to compute the
+characteristic impedance and velocity factor of the cable.
 
-From transmission line theory it is known, that the characteristic impedance of a lossless transmission line is
+From transmission line theory it is known, that the characteristic impedance
+of a lossless transmission line is
 
 .. math::
    Z = \\sqrt{\\frac{L}{C}},
@@ -16,15 +21,20 @@ while the phase velocity is
 .. math::
    v = \\frac{1}{\\sqrt{L C}},
 
-where :math:`L` is the series inductance per unit length of the transmission line and :math:`C` is the parallel capacitance per unit length of the transmission line.
+where :math:`L` is the series inductance per unit length of the transmission
+line and :math:`C` is the parallel capacitance per unit length of the
+transmission line.
 
-Further, the phase velocity relative to the speed of light is called the velocity factor of the transmission line.
+Further, the phase velocity relative to the speed of light is called the
+velocity factor of the transmission line.
 
 RG316
 -----
 
-A piece of coaxial cable is composed of an inner conductor, which is surrounded by a dielectric insulator.
-The dielectric insulator in turn is surrounded by an outer conductor. Finally, the outer conductor is surrounded by an outer insulator.
+A piece of coaxial cable is composed of an inner conductor, which is
+surrounded by a dielectric insulator.
+The dielectric insulator in turn is surrounded by an outer conductor. Finally,
+the outer conductor is surrounded by an outer insulator.
 
 For an RG316 cable, the dimensions and materials of the components are
 
@@ -33,7 +43,8 @@ For an RG316 cable, the dimensions and materials of the components are
 - Outer conductor: OD 1.98mm, silver plated copper
 - Outer insulator: OD 2.48mm, PEF
 
-RG316 has a nominal characteristic impedance of 50 ohms and a velocity factor of 0.69.
+RG316 has a nominal characteristic impedance of 50 ohms and a velocity factor
+of 0.69.
 
 Inductance
 ----------
@@ -52,29 +63,54 @@ and the constitutive relation
 
 where :math:`\\mu` is the permability of the medium.
 
-Since :math:`\\boldsymbol{B}` is divergence free, it can be written in terms of a vector potential :math:`\\boldsymbol{A}` as
+Since :math:`\\boldsymbol{B}` is divergence free, it can be written in terms
+of a vector potential :math:`\\boldsymbol{A}` as
 
 .. math::
    \\boldsymbol{B} = \\nabla \\times \\boldsymbol{A}.
 
-Thus we have the strong form for the vector potential :math:`\\boldsymbol{A}` as
[…]
