### 2020-07-11 d0568a7
More fixing of flake8 complaints
