### 2020-07-11 d9ed172
Refactor Dofnum into a separate class
