+++ skfem/assembly/basis/basis.py
@@ -6,6 +6,7 @@ import numpy as np
 from numpy import ndarray
 
 from skfem.assembly.dofs import Dofs
+from skfem.assembly.dofnum import Dofnum
 from skfem.element.discrete_field import DiscreteField
 from skfem.element.element_composite import ElementComposite
 
@@ -25,14 +26,15 @@ class Basis:
 
     N: int = 0
     dofnames: List[str] = []
+    tind: ndarray = None
 
     def __init__(self, mesh, elem, mapping=None):
 
         self.mapping = mesh.mapping() if mapping is None else mapping
 
-        self._build_dofnum(mesh, elem)
+        self.dofnum = Dofnum(mesh, elem)
 
-        if not isinstance(mesh, elem.mesh_type):
+        if mesh.refdom != elem.refdom:
             raise ValueError("Incompatible Mesh and Element.")
 
         # human readable names
@@ -60,71 +62,31 @@ class Basis:
         self.refdom = mesh.refdom
         self.brefdom = mesh.brefdom
 
-    def _build_dofnum(self, mesh, element):
-        """Build global degree-of-freedom numbering."""
-        # vertex dofs
-        self.nodal_dofs = np.reshape(
-            np.arange(element.nodal_dofs * mesh.p.shape[1], dtype=np.int64),
-            (element.nodal_dofs, mesh.p.shape[1]),
-            order='F')
-        offset = element.nodal_dofs * mesh.p.shape[1]
+    @property
+    def nodal_dofs(self):
+        return self.dofnum.nodal_dofs
 
-        # edge dofs
-        if mesh.dim() == 3:
-            self.edge_dofs = np.reshape(
-                np.arange(element.edge_dofs * mesh.edges.shape[1],
-                          dtype=np.int64),
-                (element.edge_dofs, mesh.edges.shape[1]),
-                order='F') + offset
-            offset = offset + element.edge_dofs * mesh.edges.shape[1]
-        else:
-            self.edge_dofs = np.empty((0, 0))
+    @property
+    def facet_dofs(self):
+        return self.dofnum.facet_dofs
 
-        # facet dofs
-        self.facet_dofs = np.reshape(
-            np.arange(element.facet_dofs * mesh.facets.shape[1],
-                      dtype=np.int64),
-            (element.facet_dofs, mesh.facets.shape[1]),
-            order='F') + offset
-        offset = offset + element.facet_dofs * mesh.facets.shape[1]
+    @property
+    def edge_dofs(self):
+        return self.dofnum.edge_dofs
 
-        # interior dofs
-        self.interior_dofs = np.reshape(
-            np.arange(element.interior_dofs * mesh.t.shape[1], dtype=np.int64),
-            (element.interior_dofs, mesh.t.shape[1]),
-            order='F') + offset
+    @property
+    def N(self):
+        return self.dofnum.N
 
-        # global numbering
-        self.element_dofs = np.zeros((0, mesh.t.shape[1]), dtype=np.int64)
+    @property
+    def interior_dofs(self):
+        return self.dofnum.interior_dofs
 
-        # nodal dofs
-        for itr in range(mesh.t.shape[0]):
-            self.element_dofs = np.vstack((
-                self.element_dofs,
-                self.nodal_dofs[:, mesh.t[itr]]
-            ))
-
-        # edge dofs
-        if mesh.dim() == 3:
-            for itr in range(mesh.t2e.shape[0]):
-                self.element_dofs = np.vstack((
-                    self.element_dofs,
-                    self.edge_dofs[:, mesh.t2e[itr]]
-                ))
-
-        # facet dofs
-        if mesh.dim() >= 2:
-            for itr in range(mesh.t2f.shape[0]):
-                self.element_dofs = np.vstack((
[…]
