### 2020-07-11 893fe60
Fix warnings related to docstring contents
