+++ docs/examples/ex34.py
@@ -1,4 +1,4 @@
-"""Euler-Bernoulli beam.
+r"""Euler-Bernoulli beam.
 
 This example solves the Euler-Bernoulli beam equation
 
