### 2020-07-11 1760b91
Clarify some of the types
