### 2020-07-11 c4078c4
Change test ex35 reference value to match value computed on newer gmsh mesh, and increase testing tolerance
