### 2020-07-12 #2369 @kinnala
From above:

> Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the target repository and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below.

Feel free to use our issue tracker if you feel it's appropriate.
