### 2020-07-12 062dc9c
Fix type of returned edge indices
