### 2020-07-12 d1aba7f
Update license and move exception to a separate file
