+++ LICENSE.md
@@ -1,6 +1,6 @@
 The main source code of scikit-fem is distributed with the following license.
 
-> Copyright 2018 Tom Gustafsson, Geordie McBain
+> Copyright 2018-2020 scikit-fem developers
 > 
 > Redistribution and use in source and binary forms, with or without
 > modification, are permitted provided that the following conditions are met:
@@ -27,6 +27,8 @@ The main source code of scikit-fem is distributed with the following license.
 > LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT
 > OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-Some functions or classes included within the source code distribution may
-have a different license. If this is the case, it is mentioned explicitly in
-the docstring comment.
+The above license applies to all files not included in the following list of
+exceptions.  The licenses of the following files are included in the docstring
+comment of the respective module.
+
+- skfem/zoo/tet_tensor.py, Copyright (c) 2016-2018 Nico Schlömer
