+++ skfem/zoo/tet_tensor.py
@@ -59,7 +59,7 @@ def build(x, y, z):
     elems0[0::2, 1::2, 0::2, 2] -= 1
     elems0[0::2, 0::2, 1::2, 2] -= 1
     elems0[1::2, 1::2, 1::2, 2] -= 1
-        
+
     elems0[1::2, 0::2, 0::2, 3] += 1
     elems0[0::2, 1::2, 0::2, 3] += 1
     elems0[0::2, 0::2, 1::2, 3] += 1
