### 2020-07-12 aa026e2
Include exceptions also in LICENSE.md
