### 2020-07-12 ea1e72b
Move the functionality of Dofnum into Dofs
