+++ skfem/assembly/dofs.py
@@ -1,86 +1,287 @@
 import warnings
 from typing import NamedTuple, Dict, Union, List
+from copy import deepcopy
 
 import numpy as np
 from numpy import ndarray
 
+from skfem.element import Element
+from skfem.mesh import Mesh
 
-class Dofs(NamedTuple):
-    """An object containing a subset of degree-of-freedom indices."""
 
-    nodal: Dict[str, ndarray] = {}
-    facet: Dict[str, ndarray] = {}
-    edge: Dict[str, ndarray] = {}
-    interior: Dict[str, ndarray] = {}
+class DofsView:
 
-    def all(self, key: Union[List[str], str] = None):
-        """Return an array consisting of all degrees-of-freedom.
+    nodal_ix: ndarray = None
+    edge_ix: ndarray = None
+    facet_ix: ndarray = None
+    interior_ix: ndarray = None
 
-        Parameters
-        ----------
-        key
-            Optionally, return all degrees-of-freedom corresponding to a
-            specific 'key'.
+    def __init__(self,
+                 obj,
+                 nodal_ix=None,
+                 edge_ix=None,
+                 facet_ix=None,
+                 interior_ix=None):
+        self.nodal_ix = np.empty((0,), dtype=np.int64)\
+            if nodal_ix is None else nodal_ix
+        self.edge_ix = np.empty((0,), dtype=np.int64)\
+            if edge_ix is None else edge_ix
+        self.facet_ix = np.empty((0,), dtype=np.int64)\
+            if facet_ix is None else facet_ix
+        self.interior_ix = np.empty((0,), dtype=np.int64)\
+            if interior_ix is None else interior_ix
+        self._obj = obj
 
-        Returns
-        -------
-        ndarray
-            A list of degree-of-freedom indices.
-
-        """
-        if isinstance(key, list):
-            return np.concatenate(tuple([self.all(k) for k in key]))
-
-        if key is None:
-            nodal = np.array([self.nodal[k] for k in self.nodal])
-            facet = np.array([self.facet[k] for k in self.facet])
-            edge = np.array([self.edge[k] for k in self.edge])
-            interior = np.array([self.interior[k] for k in self.interior])
+    def all(self, keep_dofnames=None):
+        if keep_dofnames is not None:
+            i1, i2, i3, i4 = self._dofnames_to_ix(keep_dofnames)
         else:
-            if key in self.nodal:
-                nodal = self.nodal[key]
-            else:
-                nodal = np.array([])
-            if key in self.facet:
-                facet = self.facet[key]
-            else:
-                facet = np.array([])
-            if key in self.edge:
-                edge = self.edge[key]
-            else:
-                edge = np.array([])
-            if key in self.interior:
-                interior = self.interior[key]
-            else:
-                interior = np.array([])
+            i1, i2, i3, i4 = 4 * (slice(None),)
+        return np.unique(
+            np.concatenate(
+                (
+                    self._obj._nodal_dofs[i1][:, self.nodal_ix].flatten(),
+                    self._obj._facet_dofs[i2][:, self.facet_ix].flatten(),
+                    self._obj._edge_dofs[i3][:, self.edge_ix].flatten(),
+                    self._obj._interior_dofs[i4][:, self.interior_ix].flatten()
+                )
+            )
+        )
 
-        output = np.concatenate((
-            nodal.flatten(),
-            facet.flatten(),
-            edge.flatten(),
-            interior.flatten()
-        )).astype(np.int)
+    @property
+    def nodal(self):
+        return self._by_name(self._nodal_dofs,
+                             ix=self.nodal_ix)
 
-        if len(output) == 0:
[…]
