### 2020-07-12 3043eb8
Fix flake8 issues
