### 2020-07-12 kinnala/scikit-fem#422 @AnjaliSandip
JOSS Review/Comments
Comments:
1.	Summary:
“The user loads a computational mesh, picks suitable basis functions, and provides the PDE’s weak formulation. This results in sparse matrices and vectors compatible with the SciPy (Virtanen et al., 2020) ecosystem.”
For someone who does not have a background in finite elements, they may not understand what  basis functions or weak formulation means.

2.	Performance:
The performance claim in vague: “Despite being fully interpreted, the code has a reasonably good performance.”
If the author could be more specific (quantifiable), it would help the reviewer in making a judgement.
