### 2020-07-12 #2369 @AnjaliSandip
I have posted the comments here:  https://github.com/kinnala/scikit-fem/issues/422
