### 2020-07-13 48b4202
Make variables private and remove unnecessary ones
