### 2020-07-13 6bcf1f3
Add changelog entry and fix flake8
