+++ README.md
@@ -111,8 +111,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Unreleased
 
 #### Added
-- `Dofs.__or__`, for merging degree-of-freedom sets (i.e. `Dofs` objects) using
-  `|` operator
+- `Dofs.__or__` and `Dofs.__add__`, for merging degree-of-freedom sets
+  (i.e. `Dofs` objects) using `|` and `+` operators
+  
+#### Removed
+- `Basis.find_dofs` no more accepts the keyword argument `skip`
 
 ### [1.2.0] - 2020-07-07
 
