### 2020-07-14 4af106d
Add name for the bubble DOF
