+++ skfem/element/element_tet/element_tet_mini.py
@@ -9,7 +9,7 @@ class ElementTetMini(ElementH1):
     interior_dofs = 1
     dim = 3
     maxdeg = 1 + dim
-    dofnames = ['u']
+    dofnames = ['u', 'NA']
     doflocs = np.array([[0., 0., 0.],
                         [1., 0., 0.],
                         [0., 1., 0.],
