### 2020-07-14 6222959
Reimplement DOF skipping
