+++ tests/test_basis.py
@@ -105,3 +105,16 @@ class TestDofsMerge(TestCase):
         D4 = basis.get_dofs(lambda x: x[1] == 0)
         assert_allclose(D1 | D2 | D3 | D4,
                         basis.get_dofs())
+        assert_allclose(D1 + D2 + D3 + D4,
+                        basis.get_dofs())
+
+
+class TestDofsSubset(TestCase):
+
+    def runTest(self):
+        m = MeshTri()
+        basis = InteriorBasis(m, ElementTriP2())
+        dofs = basis.get_dofs()
+
+        self.assertEqual(len(dofs.nodal['u']), 4)
+        self.assertEqual(len(dofs.facet['u']), 4)
