+++ skfem/assembly/dofs.py
@@ -6,42 +6,117 @@ from skfem.mesh import Mesh
 
 
 class DofsView:
-    """A proxy to a subset of :class:`skfem.assembly.Dofs`."""
+    """A subset of :class:`skfem.assembly.Dofs`."""
 
     _nodal_ix: ndarray = None
     _edge_ix: ndarray = None
     _facet_ix: ndarray = None
     _interior_ix: ndarray = None
 
+    _nodal_rows = slice(None)
+    _edge_rows = slice(None)
+    _facet_rows = slice(None)
+    _interior_rows = slice(None)
+
     def __init__(self,
                  obj,
                  nodal_ix=None,
-                 edge_ix=None,
                  facet_ix=None,
-                 interior_ix=None):
-        self._nodal_ix = np.empty((0,), dtype=np.int64)\
+                 edge_ix=None,
+                 interior_ix=None,
+                 nodal_rows=None,
+                 facet_rows=None,
+                 edge_rows=None,
+                 interior_rows=None):
+        self._nodal_ix = slice(None)\
             if nodal_ix is None else nodal_ix
-        self._edge_ix = np.empty((0,), dtype=np.int64)\
-            if edge_ix is None else edge_ix
-        self._facet_ix = np.empty((0,), dtype=np.int64)\
+        self._facet_ix = slice(None)\
             if facet_ix is None else facet_ix
-        self._interior_ix = np.empty((0,), dtype=np.int64)\
+        self._edge_ix = slice(None)\
+            if edge_ix is None else edge_ix
+        self._interior_ix = slice(None)\
             if interior_ix is None else interior_ix
+        self._nodal_rows = slice(None)\
+            if nodal_rows is None else nodal_rows
+        self._facet_rows = slice(None)\
+            if facet_rows is None else facet_rows
+        self._edge_rows = slice(None)\
+            if edge_rows is None else edge_rows
+        self._interior_rows = slice(None)\
+            if interior_rows is None else interior_rows
         self._obj = obj
 
-    def all(self, keep_dofnames=None):
-        if keep_dofnames is not None:
-            i1, i2, i3, i4 = self._dofnames_to_ix(keep_dofnames)
-        else:
-            i1, i2, i3, i4 = 4 * (slice(None),)
+    def flatten(self):
         return np.unique(
-            np.concatenate(
-                (self._obj._nodal_dofs[i1][:, self._nodal_ix].flatten(),
-                 self._obj._facet_dofs[i2][:, self._facet_ix].flatten(),
-                 self._obj._edge_dofs[i3][:, self._edge_ix].flatten(),
-                 self._obj._interior_dofs[i4][:, self._interior_ix].flatten())
+            np.concatenate((
+                (self._obj
+                 ._nodal_dofs[self._nodal_rows][:, self._nodal_ix]
+                 .flatten()),
+                (self._obj
+                 ._facet_dofs[self._facet_rows][:, self._facet_ix]
+                 .flatten()),
+                (self._obj
+                 ._edge_dofs[self._edge_rows][:, self._edge_ix]
+                 .flatten()),
+                (self._obj
+                 ._interior_dofs[self._interior_rows][:, self._interior_ix]
+                 .flatten())
+            ))
+        )
+
+    def _intersect(self, a, b):
+        if isinstance(a, slice):
+            if a.start == 0:
+                return a
+            return b
+        if isinstance(b, slice):
+            if b.start == 0:
+                return b
+            return a
+        return np.intersect1d(a, b)
+
+    def _intersect_tuples(self, a, b):
+        return tuple([self._intersect(a[i], b[i]) for i in range(len(a))])
+
+    def keep(self, dofnames):
+        return DofsView(
+            self._obj,
[…]
