### 2020-07-14 4bd7f58
Revert back the skipping test
