+++ tests/test_basis.py
@@ -41,13 +41,11 @@ class TestCompositeSplitting(TestCase):
 
         S = asm(bilinf, basis)
 
-        D = basis.find_dofs()
+        D = basis.find_dofs(skip=['u^2'])
         x = basis.zeros()
         x[D['up'].all('u^1^1')] = .1
 
-        x = solve(*condense(S, basis.zeros(), x=x,
-                            D=np.concatenate([v.all(['u^1^1', 'u^2^1'])
-                                              for _, v in D.items()])))
+        x = solve(*condense(S, basis.zeros(), x=x, D=D))
 
         (u, u_basis), (p, p_basis) = basis.split(x)
 
@@ -60,8 +58,7 @@ class TestCompositeSplitting(TestCase):
         self.assertTrue(isinstance(U.value, np.ndarray))
         self.assertTrue(isinstance(P.value, np.ndarray))
 
-        self.assertTrue((basis.doflocs[:, D['up'].all(['u^1^1',
-                                                       'u^2^1'])][1] == 1.).all())
+        self.assertTrue((basis.doflocs[:, D['up'].all()][1] == 1.).all())
 
 
 class TestFacetExpansion(TestCase):
