### 2020-07-14 eca8cd8
Add separate test file for Dofs
