+++ skfem/assembly/dofs.py
@@ -1,4 +1,4 @@
-from typing import Union
+from typing import Union, NamedTuple, Any
 
 import numpy as np
 from numpy import ndarray
@@ -7,61 +7,33 @@ from skfem.element import Element
 from skfem.mesh import Mesh
 
 
-class DofsView:
+class DofsView(NamedTuple):
     """A subset of :class:`skfem.assembly.Dofs`."""
 
-    _nodal_ix: Union[ndarray, slice] = None
-    _facet_ix: Union[ndarray, slice] = None
-    _edge_ix: Union[ndarray, slice] = None
-    _interior_ix: Union[ndarray, slice] = None
-
-    _nodal_rows: Union[ndarray, slice] = None
-    _facet_rows: Union[ndarray, slice] = None
-    _edge_rows: Union[ndarray, slice] = None
-    _interior_rows: Union[ndarray, slice] = None
-
-    def __init__(self,
-                 obj,
-                 nodal_ix=None,
-                 facet_ix=None,
-                 edge_ix=None,
-                 interior_ix=None,
-                 nodal_rows=None,
-                 facet_rows=None,
-                 edge_rows=None,
-                 interior_rows=None):
-        self._nodal_ix = slice(None)\
-            if nodal_ix is None else nodal_ix
-        self._facet_ix = slice(None)\
-            if facet_ix is None else facet_ix
-        self._edge_ix = slice(None)\
-            if edge_ix is None else edge_ix
-        self._interior_ix = slice(None)\
-            if interior_ix is None else interior_ix
-        self._nodal_rows = slice(None)\
-            if nodal_rows is None else nodal_rows
-        self._facet_rows = slice(None)\
-            if facet_rows is None else facet_rows
-        self._edge_rows = slice(None)\
-            if edge_rows is None else edge_rows
-        self._interior_rows = slice(None)\
-            if interior_rows is None else interior_rows
-        self._obj = obj
+    obj: Any = None
+    nodal_ix: Union[ndarray, slice] = slice(None)
+    facet_ix: Union[ndarray, slice] = slice(None)
+    edge_ix: Union[ndarray, slice] = slice(None)
+    interior_ix: Union[ndarray, slice] = slice(None)
+    nodal_rows: Union[ndarray, slice] = slice(None)
+    facet_rows: Union[ndarray, slice] = slice(None)
+    edge_rows: Union[ndarray, slice] = slice(None)
+    interior_rows: Union[ndarray, slice] = slice(None)
 
     def flatten(self):
         return np.unique(
             np.concatenate((
-                (self._obj
-                 ._nodal_dofs[self._nodal_rows][:, self._nodal_ix]
+                (self.obj
+                 .nodal_dofs[self.nodal_rows][:, self.nodal_ix]
                  .flatten()),
-                (self._obj
-                 ._facet_dofs[self._facet_rows][:, self._facet_ix]
+                (self.obj
+                 .facet_dofs[self.facet_rows][:, self.facet_ix]
                  .flatten()),
-                (self._obj
-                 ._edge_dofs[self._edge_rows][:, self._edge_ix]
+                (self.obj
+                 .edge_dofs[self.edge_rows][:, self.edge_ix]
                  .flatten()),
-                (self._obj
-                 ._interior_dofs[self._interior_rows][:, self._interior_ix]
+                (self.obj
+                 .interior_dofs[self.interior_rows][:, self.interior_ix]
                  .flatten())
             ))
         )
@@ -82,32 +54,32 @@ class DofsView:
 
     def keep(self, dofnames):
         return DofsView(
-            self._obj,
-            self._nodal_ix,
-            self._facet_ix,
-            self._edge_ix,
-            self._interior_ix,
[…]
