+++ skfem/utils.py
@@ -2,8 +2,7 @@
 SciPy linear solvers."""
 
 import warnings
-from typing import Optional, Union, Tuple,\
-    Callable, Dict
+from typing import Optional, Union, Tuple, Callable, Dict
 
 import numpy as np
 import scipy.sparse as sp
@@ -76,6 +75,7 @@ def solver_eigen_scipy(**kwargs) -> EigenSolver:
 
 
 def solver_direct_scipy(**kwargs) -> LinearSolver:
+    """The default linear solver of SciPy."""
     def solver(A, b, **solve_time_kwargs):
         kwargs.update(solve_time_kwargs)
         return spl.spsolve(A, b, **kwargs)
@@ -186,9 +186,9 @@ def _flatten_dofs(S: DofsCollection) -> ndarray:
         if isinstance(S, ndarray):
             return S
         elif isinstance(S, DofsView):
-            return S.all()
+            return S.flatten()
         elif isinstance(S, dict):
-            return np.unique(np.concatenate([S[key].all() for key in S]))
+            return np.unique(np.concatenate([S[key].flatten() for key in S]))
         raise NotImplementedError("Unable to flatten the given set of DOF's.")
 
 
@@ -275,9 +275,9 @@ def rcm(A: spmatrix,
 def adaptive_theta(est, theta=0.5, max=None):
     """For choosing which elements to refine in an adaptive strategy."""
     if max is None:
-        return np.nonzero(theta*np.max(est) < est)[0]
+        return np.nonzero(theta * np.max(est) < est)[0]
     else:
-        return np.nonzero(theta*max < est)[0]
+        return np.nonzero(theta * max < est)[0]
 
 
 def project(fun,
@@ -341,8 +341,10 @@ def project(fun,
 
 # for backwards compatibility
 def L2_projection(a, b, c=None):
+    """Superseded by :func:`skfem.utils.project`."""
     return project(a, basis_to=b, I=c)
 
 
 def derivative(a, b, c, d=0):
+    """Superseded by :func:`skfem.utils.project`."""
     return project(a, basis_from=b, basis_to=c, diff=d)
