### 2020-07-14 ff98d8c
More tests and clearer code
