+++ tests/test_dofs.py
@@ -43,6 +43,15 @@ class TestDofsNodalSubsets(TestCase):
         self.assertEqual(len(all_dofs.keep('u').nodal), 1)
         self.assertTrue('u' in all_dofs.keep('u').nodal)
 
+        self.assertEqual(len(all_dofs.keep('u_n').facet), 1)
+        self.assertEqual(len(all_dofs.drop('u').facet), 1)
+
+        all_dofs = basis.dofs.get_facet_dofs(
+            m.facets_satisfying(lambda x: 1))
+
+        self.assertEqual(len(all_dofs.keep('u_n').facet), 1)
+        self.assertEqual(len(all_dofs.drop('u').facet), 1)
+
 
 class TestDofsMerge(TestCase):
 
