### 2020-07-15 kinnala/scikit-fem#426 @thelfer
Comments on your JOSS paper review
This documents is part of the review of the paper submitted to the Journal of Open Source Software. See <https://github.com/openjournals/joss-reviews/issues/2369>.

The paper describes the `scikit-fem` `python` package which is dedicated to the assembly of the matrices and vectors resulting from the finite element discretization of partial differential equations. More precisely, the library allows the user to focus on the weak form of the differential equation and the choice of the finite element library (several families are provided).

The paper illustrates the library with three examples, grouped in a single figure:

- a linear elastic contact problem.
- a H(curl)-conforming model problem.
- a resolution of the Navier–Stokes equations for different Reynolds numbers using Taylor–Hood element.

# General remarks and questions

- The paper is easy to read. I don't have particular remarks on the quality of english used.
- The paper is very short. I understand that the task is to library is very narrow. However, many things are missing to fully understand the interest and the limits of the library. For example:
  - The library do provide a way to handle Dirichlet boundary conditions by elimination. This shall be at least mentioned. I imagine that the treatment of more complex conditions, such as multiple points constraints in mechanics, have to be treated by hand after the assembly, isn't it ?
  - Following the previous remark, it is not clear how the library could be used in this way. For example, is the mesh connectivity clearly described (how the unknowns are numbered) ?
  - The authors shall describe how the library could be used in non linear cases. In particular, the contact problem given in the first example is indeed non linear.
  - Can non uniform coefficients be handled ? To make my question clearer, would I be able to treat:
    - a linear elastic problem with a non uniform Young Modulus.
    - a heat transfer equation with a non uniform heat source Those non uniform coefficients are dealt with in `UFL` using the notion of functions spaces. Is there anything equivalent in `scikit-fem` ?
- The authors claims that the assembly, while made in pure `python`, is  "fast enough" ? This shall be made more explicit. Do the authors mean that the assembly takes a negligible time compared to the resolution of the resulting linear system ?
- The authors claims that more than 30 examples illustrate the use of the library. I did not find any page that summarises those examples. It would be very useful for new users. I mean something similar to what the `MFEM` developers provide: <https://mfem.org/examples/>
- Would it work in MPI-like processes ? Do you plan any extension of the package in that matter or is it out of its scope ?.

# Specific comments

## Quadrature points

`UFL` offers functional spaces at quadrature points. This functional spaces are relevant when dealing with non linear constitutive equations in solid mechanics. Is there anything similar in `scikit-fem` ?

## Figures

With which package have the figures been generated ?

## Second example

The description of the second example is very sparse.

## Dependencies

Some examples depends on external packages which are not installed with `scikit-fem`. An appropriate list of dependencies would be appreciated.

[review.pdf](https://github.com/kinnala/scikit-fem/files/4923631/review.pdf)
