### 2020-07-15 #2369 @thelfer
@meg-simula Sorry for the time I spent on the review.

@kinnala I posted some comments and general remarks on the paper here: https://github.com/kinnala/scikit-fem/issues/426.
