### 2020-07-15 45fb6d3
Add docstrings
