### 2020-07-15 kinnala/scikit-fem#426 @kinnala
Thanks for valuable feedback. Some clarifying questions:

> Would it work in MPI-like processes ? Do you plan any extension of the package in that matter or is it out of its scope ?.

It's currently out of scope. Do you ask this to be discussed in the paper?

> The description of the second example is very sparse.

Do you mean the following example?
https://kinnala.github.io/scikit-fem-docs/example2.html

> UFL offers functional spaces at quadrature points. This functional spaces are relevant when dealing with non linear constitutive equations in solid mechanics. Is there anything similar in scikit-fem ?

I'm not sure I recognize the concept but I have used scikit-fem library to solve, e.g., nonlinear (Neohookean) elasticity. Is this (functional spaces at quadrature points) something related to plasticity?
