### 2020-07-16 kinnala/scikit-fem#426 @thelfer
> > Would it work in MPI-like processes ? Do you plan any extension of the package in that matter or is it out of its scope ?.
>
> It's currently out of scope. Do you ask this to be discussed in the paper?

It would be nice to have a better view of the limitations and perspectives of the package. But that's only a point of discussion among (many) others.
