### 2020-07-16 kinnala/scikit-fem#426 @thelfer
> > The description of the second example is very sparse.
>
> Do you mean the following example?
> https://kinnala.github.io/scikit-fem-docs/example2.html

No. I mean the second example of the paper: "a H(curl)-conforming model problem."
