### 2020-07-16 kinnala/scikit-fem#426 @thelfer
> > UFL offers functional spaces at quadrature points. This functional spaces are relevant when dealing with non linear constitutive equations in solid mechanics. Is there anything similar in scikit-fem ?
>
> I'm not sure I recognize the concept but I have used scikit-fem library to solve, e.g., nonlinear (Neohookean) elasticity. Is this (functional spaces at quadrature points) something related to plasticity?

In a way, it is related to plasticity. More precisely, I am the main author of `MFront` (http://tfel.sourceforge.net/) which is a set of domain specific languages dedicated to advanced constitutive equations (such as plasticity). This explains my interest on this topic.

Advanced constitutive equations describes the evolution of internal state variables (elastic strain, dislocation densities, equivalent plastic strain, back-strains, damage, etc....). One the values of the internal state variables are known, the stresss can be deduced, so the knowledge of those values are required at the quadrature points.
