### 2020-07-16 kinnala/scikit-fem#426 @kinnala
> Advanced constitutive equations describes the evolution of internal state variables (elastic strain, dislocation densities, equivalent plastic strain, back-strains, damage, etc....). One the values of the internal state variables are known, the stresss can be deduced, so the knowledge of those values are required at the quadrature points.

I think I found a discussion of the concept in Chapter 26 of https://launchpadlibrarian.net/83776282/fenics-book-2011-10-27-final.pdf

I'll try to read it through and see if there is anything comparable in scikit-fem.
