### 2020-07-16 d2b8b04
Start work on gallery of examples
