+++ docs/example1.rst
@@ -1,9 +1,0 @@
-.. note::
-
-   More examples can be found `in the source code distribution <https://github.com/kinnala/scikit-fem/tree/master/docs/examples>`_.
-
-==========================
- Example: Kirchhoff plate
-==========================
-
-.. automodule:: docs.examples.ex02
