+++ docs/overview.rst
@@ -1,6 +1,6 @@
-==============================
- An overview of main features
-==============================
+===========================
+ Overview of main features
+===========================
 
 Module: skfem.mesh
 ==================
