### 2020-07-16 49eaf8d
Add more examples to the gallery
