### 2020-07-17 053e0b4
Remove unused import of Optional
