+++ skfem/assembly/form/bilinear_form.py
@@ -42,6 +42,7 @@ class BilinearForm(Form):
     def assemble(self,
                  u: Basis,
                  v: Optional[Basis] = None,
+                 dtype = np.float,
                  **kwargs) -> Any:
 
         if v is None:
@@ -56,7 +57,7 @@ class BilinearForm(Form):
 
         # initialize COO data structures
         sz = u.Nbfun * v.Nbfun * nt
-        data = np.zeros(sz)
+        data = np.zeros(sz, dtype=dtype)
         rows = np.zeros(sz)
         cols = np.zeros(sz)
 
