+++ skfem/assembly/form/bilinear_form.py
@@ -42,7 +42,7 @@ class BilinearForm(Form):
     def assemble(self,
                  u: Basis,
                  v: Optional[Basis] = None,
-                 dtype = np.float,
+                 dtype=np.float,
                  **kwargs) -> Any:
 
         if v is None:
