+++ .travis.yml
@@ -14,7 +14,7 @@ install:
   - tar zvxf gmsh-4.5.2-Linux64.tgz
   - sudo cp gmsh-4.5.2-Linux64/bin/gmsh /usr/bin
   - cd ..
-  - wget https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh;
+  - wget -nc https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh;
   - bash miniconda.sh -b -p $HOME/miniconda
   - export PATH="$HOME/miniconda/bin:$PATH"
   - hash -r
