### 2020-07-17 bbd713d
Update miniconda version
