### 2020-07-17 eda1125
Change to new miniconda url
