### 2020-07-18 c17495d
Add images for the rest of the examples
