### 2020-07-19 68c506b
Optional dtype in form decorator
