### 2020-07-20 cb59d07
Add rest of the descriptions
