+++ docs/listofexamples.rst
@@ -44,7 +44,7 @@ of a cantilever beam.
 Example 4: Linearized contact problem
 =====================================
 
-This example solves a single interation of the linearized contact problem
+This example solves a single interation of the contact problem
 between two elastic bodies using the Nitsche's method.
 Triangular and quadrilateral second-order elements are used
 in the discretization of the two elastic bodies.
@@ -236,15 +236,7 @@ Example 18: Stokes equations
 
 This example solves for the creeping flow problem in the primitive variables,
 i.e. velocity and pressure instead of the stream-function.  These are governed
-by the Stokes momentum
-
-.. math::
-   - \nu\Delta\boldsymbol{u} + \rho^{-1}\nabla p = \boldsymbol{f}
-
-and continuity equation
-
-.. math::
-   \nabla\cdot\boldsymbol{u} = 0.
+by the Stokes momentum :math:`- \nu\Delta\boldsymbol{u} + \rho^{-1}\nabla p = \boldsymbol{f}` and the continuity equation :math:`\nabla\cdot\boldsymbol{u} = 0`.
 
 .. note::
    This example requires the external package `dmsh <https://pypi.org/project/dmsh/>`_.
@@ -254,19 +246,9 @@ and continuity equation
 `Source code <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex18.py>`_
 
 Example 19: Heat equation
-=============
+=========================
 
-This example solves the heat equation
-
-.. math::
-    \frac{\partial T}{\partial t} = \kappa\Delta T
-    
-in the domain :math:`|x|<w_0` and :math:`|y|<w_1` with the initial value
-
-.. math::
-    T(x,y,0) = \cos\frac{\pi x}{2w_0}\cos\frac{\pi y}{2w_1}
-
-using the generalized trapezoidal
+This example solves the heat equation :math:`\frac{\partial T}{\partial t} = \kappa\Delta T` in the domain :math:`|x|<w_0` and :math:`|y|<w_1` with the initial value :math:`T_0(x,y) = \cos\frac{\pi x}{2w_0}\cos\frac{\pi y}{2w_1}` using the generalized trapezoidal
 rule ("theta method").
 
 .. note::
@@ -276,84 +258,233 @@ rule ("theta method").
 
 `Source code <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex19.py>`_
 
-Example 20: -
-=============
+Example 20: Creeping flow via stream-function
+=============================================
+
+This example solves the creeping flow problem via the stream-function
+formulation.
+The stream-function :math:`\psi` for two-dimensional creeping flow is
+governed by the biharmonic equation :math:`\nu \Delta^2\psi = \mathrm{rot}\,\boldsymbol{f}` where :math:`\nu` is the kinematic viscosity (assumed constant),
+:math:`\boldsymbol{f}` the volumetric body-force, and :math:`\mathrm{rot}\,\boldsymbol{f} =
+\partial f_y/\partial x - \partial f_x/\partial y`.  The boundary
+conditions at a wall are that :math:`\psi` is constant (the wall is
+impermeable) and that the normal component of its gradient vanishes (no
+slip)
+
+.. note::
+   This example requires the external package `pygmsh <https://pypi.org/project/pygmsh/>`_.
 
 .. figure:: https://user-images.githubusercontent.com/973268/87778910-9745a580-c834-11ea-8277-62d58a7fe7b8.png
 
-Example 21: -
-=============
+`Source code <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex20.py>`_
+
+Example 21: Structural vibration
+================================
+
+This example demonstrates the solution of a three-dimensional vector-valued
[…]
