### 2020-07-20 56dc6ed
Remove examples 5 and 6 from gallery temporarily
