### 2020-07-20 d36e0a5
Fix some Sphinx warnings
