+++ docs/listofexamples.rst
@@ -14,7 +14,7 @@ triangular elements.
 
 .. figure:: https://user-images.githubusercontent.com/973268/87638021-c3d1c280-c74b-11ea-9859-dd82555747f5.png
 
-`Source code <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex01.py>`_
+`Source code of Example 1 <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex01.py>`_
            
 Example 2: Kirchhoff plate bending problem
 ==========================================
@@ -26,7 +26,7 @@ combination of clamped, simply supported and free boundary conditions.
 
 .. figure:: https://user-images.githubusercontent.com/973268/87659951-f50bbc00-c766-11ea-8c0e-7de0e9e83714.png
 
-`Source code <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex02.py>`_
+`Source code of Example 2 <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex02.py>`_
 
 Example 3: Linear elastic eigenvalue problem
 ============================================
@@ -39,7 +39,7 @@ of a cantilever beam.
 
 .. figure:: https://user-images.githubusercontent.com/973268/87661134-cbec2b00-c768-11ea-81bc-f5455df7cc33.png
 
-`Source code <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex03.py>`_
+`Source code of Example 3 <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex03.py>`_
 
 Example 4: Linearized contact problem
 =====================================
@@ -55,7 +55,7 @@ in the discretization of the two elastic bodies.
 
 .. figure:: https://user-images.githubusercontent.com/973268/87661313-1372b700-c769-11ea-89ee-db144986a25a.png
 
-`Source code <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex04.py>`_
+`Source code of Example 4 <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex04.py>`_
 
 Example 7: Interior penalty method
 ==================================
@@ -67,7 +67,7 @@ the element edges.
 
 .. figure:: https://user-images.githubusercontent.com/973268/87662192-80d31780-c76a-11ea-9291-2d11920bc098.png
 
-`Source code <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex07.py>`_
+`Source code of Example 7 <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex07.py>`_
 
 Example 8: Argyris basis functions
 ==================================
@@ -78,7 +78,7 @@ This element can be used in the conforming discretization of biharmonic problems
 
 .. figure:: https://user-images.githubusercontent.com/973268/87662432-e0c9be00-c76a-11ea-85b9-711c6b34791e.png
 
-`Source code <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex08.py>`_
+`Source code of Example 8 <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex08.py>`_
 
 Example 9: Three-dimensional Poisson equation
 =============================================
@@ -95,7 +95,7 @@ conjugate gradient method.  The figure was created using `Paraview
 .. figure:: https://user-images.githubusercontent.com/973268/87681574-7a06cd80-c787-11ea-8cfd-6ff5079e752c.png
    :width: 500px
 
-`Source code <https://github.com/kinnala/scikit-fem/blob/master/docs/examples/ex09.py>`_
[…]
