### 2020-07-20 kinnala/scikit-fem#426 @kinnala
There is now a gallery of examples at https://scikit-fem.readthedocs.io/en/latest/listofexamples.html .
