### 2020-07-22 ac21922
ex10 with JFNK, as discussed in #119 #439
