+++ skfem/assembly/__init__.py
@@ -26,7 +26,7 @@ assembly. The basic workflow of assembly is the following:
 >>> b = asm(form_l, basis)
 >>> A
 <4x4 sparse matrix of type '<class 'numpy.float64'>'
-	with 14 stored elements in Compressed Sparse Row format>
+       with 14 stored elements in Compressed Sparse Row format>
 >>> b
 array([0.0162037 , 0.15046296, 0.06712963, 0.09953704])
 
