### 2020-07-23 96484e3
More work documenting boundary conditions
