+++ docs/bcs.rst
@@ -10,12 +10,12 @@ Using the techniques described in :ref:`overview`, one obtains the linear system
 where :math:`A` corresponds to a bilinear form and :math:`b` corresponds to a
 linear form.
 Many times this system has no unique solution unless the degrees-of-freedom
-:math:`x` are further constrained by imposing boundary conditions.
+(DOF's) :math:`x` are further constrained by imposing boundary conditions.
 
 Essential boundary conditions
 =============================
 
-It is possible to eliminate degrees-of-freedom from the resulting system if some
+It is possible to eliminate DOF's from the resulting system if some
 of the values are known `a priori`.  Suppose that the vector :math:`x` can be
 split as
 
@@ -65,8 +65,8 @@ corresponding to the Poisson equation :math:`-\Delta u = 1`.
    >>> A = laplace.assemble(basis)
    >>> b = unit_load.assemble(basis)
 
-The condensed system can be obtained with :func:`skfem.utils.condense`.  Below
-we provide a list of degrees-of-freedom to eliminate via the keyword argument
+The condensed system is obtained with :func:`skfem.utils.condense`.  Below
+we provide the DOF's to eliminate via the keyword argument
 ``D``.
 
 .. code-block:: python
@@ -80,17 +80,18 @@ we provide a list of degrees-of-freedom to eliminate via the keyword argument
            0., 0., 0., 0., 0., 0., 0., 0.]),
     array([ 6, 12, 15, 19, 20, 21, 22, 23, 24]))
 
-By default, the eliminated degrees-of-freedom are assumed to be zero.
-Other values can be provided through the keyword argument ``x``,
-cf. :ref:`ex14`.
+By default, the eliminated DOF's are set to zero.
+Different values can be provided through the keyword argument ``x``;
+see :ref:`ex14`.
 
 Finding degrees-of-freedom
 ==========================
 
-Often we wish to constrain degrees-of-freedom (DOF's) on a specific part of the
-domain.  Currently the main tools for finding DOF's are
+Often the goal is to constrain DOF's on a specific part of
+the boundary.  Currently the main tools for finding DOF's are
 :meth:`skfem.assembly.Basis.find_dofs` and
-:meth:`skfem.assembly.Basis.get_dofs`.
+:meth:`skfem.assembly.Basis.get_dofs`.  Let us demonstrate
+the latter with an example.
 
 .. code-block:: python
 
@@ -99,21 +100,71 @@ domain.  Currently the main tools for finding DOF's are
    >>> m.refine(2)
    >>> basis = InteriorBasis(m, ElementTriP2())
 
-Finding DOF's corresponding to the left boundary:
+We first find the set of facets belonging to the left boundary.
 
+.. code-block:: python
+
+   >>> m.facets_satisfying(lambda x: x[0] == 0.)
+   array([ 1,  5, 14, 15])
+
+Next we supply the array of facet indices to
+:meth:`skfem.assembly.Basis.get_dofs`
+
+.. code-block:: python
+
+   >>> dofs = basis.get_dofs(m.facets_satisfying(lambda x: x[0] == 0.))
+   >>> dofs.nodal
+   {'u': array([ 0,  2,  5, 10, 14])}
+   >>> dofs.facet
+   {'u': array([26, 30, 39, 40])}
+
+The keys in the above dictionaries indicate the type of the
+DOF according to the following table:
+
++-----------+---------------------------------------------------------------+
+| Key       | Description                                                   |
++===========+===============================================================+
+| ``u``     | Point value                                                   |
++-----------+---------------------------------------------------------------+
+| ``u_n``   | Normal derivative                                             |
++-----------+---------------------------------------------------------------+
+| ``u_x``   | Partial derivative w.r.t. :math:`x`                           |
++-----------+---------------------------------------------------------------+
[…]
