+++ README.md
@@ -127,9 +127,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Fixed
 - `Mesh3D.boundary_edges` was broken in case of hexahedral meshes
 - `skfem.utils.project` did not work for `ElementGlobal`
+- `ElementQuadP` did not work with `FacetBasis`
 
 #### Changed
 - `MeshQuad._splitquads` aliased as `MeshQuad.to_meshtri`: should not be private
+
+#### Deprecated
+- `skfem.utils.project` will support functions like `lambda x: x[0]`
+  instead of `lambda x, y, z: x` in the future
 
 ### [1.1.0] - 2020-05-18
 
