### 2020-07-23 31aeb26
Remove unused lines
