### 2020-07-23 3a273c5
Allow condense(..., expand=True) in project
