### 2020-07-24 f8892b1
Add information about 'w' argument
