+++ docs/forms.rst
@@ -15,13 +15,8 @@ operator :math:`-\Delta` is
 
    a(u, v) = \int_\Omega \nabla u \cdot \nabla v \,\mathrm{d}x.
 
-In order to express this in scikit-fem we pick the integrand,
-
-.. math::
-
-   a(u, v) = \int_\Omega \underbrace{\nabla u \cdot \nabla v}_{~} \,\mathrm{d}x,
-
-and write it as a Python function:
+In order to express this in scikit-fem we pick the integrand and write it as a
+Python function:
 
 .. code-block:: python
 
@@ -78,11 +73,11 @@ readable.  An alternative way to write the above form is
    def integrand(u, v, w):
        return u[1][0] * v[1][0] + u[1][1] * v[1][1]
 
-In fact, ``u`` and ``v`` are simply tuples of NumPy arrays
+In fact, ``u`` and ``v`` are more or less tuples of NumPy arrays
 with the values of the function at ``u[0]`` and the values
-of the gradient at ``u[1]``.
-In addition, they implement ``__array__`` and ``__mul__``
-so that expressions such as ``u * v`` work as expected.
+of the gradient at ``u[1]`` (and some additional magic such as
+implementing ``__array__`` and ``__mul__``
+so that expressions such as ``u * v`` work as expected).
 
 Notice how the shape of ``u[0]`` is what we expect also from the return value:
 
@@ -96,3 +91,64 @@ Notice how the shape of ``u[0]`` is what we expect also from the return value:
    array([[0.66666667, 0.16666667, 0.16666667],
           [0.66666667, 0.16666667, 0.16666667]])
 
+
+Using predefined functions in the forms
+=======================================
+
+It is sometimes necessary to use a previous solution vector in the form
+definition, e.g., when solving nonlinear problems.
+A simple fixed-point iteration for solving the nonlinear boundary
+value problem
+
+.. math::
+
+   \begin{aligned}
+      -\nabla \cdot ((u + 1)\nabla u) &= 1 \quad \text{in $\Omega$} \\
+      u &= 0 \quad \text{on $\partial \Omega$}
+   \end{aligned}
+
+would correspond to repeatedly
+finding :math:`u_{k+1} \in H^1_0(\Omega)` which satisfies
+
+.. math::
+
+   \int_\Omega (u_{k} + 1) \nabla u_{k+1} \cdot \nabla v \,\mathrm{d}x = \int_\Omega v\,\mathrm{d}x
+
+for every :math:`v \in H^1_0(\Omega)` so that the previous approximation
+:math:`u_k` is given.
+
+Defining such forms requires the use of the argument ``w``:
+
+.. code-block:: python
+
+   >>> from skfem import *
+   >>> from skfem.models.poisson import unit_load
+   >>> from skfem.helpers import grad, dot
+   >>> @BilinearForm
+   ... def bilinf(u, v, w):
+   ...     return (w.u_k + 1.) * dot(grad(u), grad(v))
+
+When performing the fixed point iteration, we provide the previous
+solution to :func:`skfem.assembly.asm` as a keyword argument:
+
+.. code-block:: python
+
+   >>> m = MeshTri()
+   >>> m.refine(3)
+   >>> basis = InteriorBasis(m, ElementTriP1())
+   >>> b = asm(unit_load, basis)
+   >>> x = 0. * b.copy()
+   >>> for itr in range(10):  # fixed point iteration
+   ...     A = asm(bilinf, basis, u_k=basis.interpolate(x))
+   ...     x = solve(*condense(A, b, I=m.interior_nodes()))
+   ...     print(x.max())
+   0.07278262867647059
+   0.07030433694174187
+   0.07036045457157739
[…]
