### 2020-07-24 85b0f2f
Add new article and information about dependencies
