### 2020-07-24 06ac3e1
Add performance benchmark
