### 2020-07-27 98eec0a
Use predefined forms in the benchmark
