### 2020-07-28 2e3c9a3
Update README with larger benchmark
