### 2020-07-28 aeb4e7d
Remove remark on the size of the matrix
