### 2020-07-28 kinnala/scikit-fem#422 @kinnala
> 2. Performance:
>    The performance claim in vague: “Despite being fully interpreted, the code has a reasonably good performance.”
>    If the author could be more specific (quantifiable), it would help the reviewer in making a judgement.

There is now a simple benchmark in the README (https://github.com/kinnala/scikit-fem). Do you find it sufficient?
