### 2020-07-28 kinnala/scikit-fem#426 @kinnala
> Some examples depends on external packages which are not installed with scikit-fem. An appropriate list of dependencies would be appreciated.

There is now an explicit discussion of installation dependencies and test dependencies (all examples are tested) in the README. Do you find it sufficient?
