### 2020-07-28 kinnala/scikit-fem#426 @kinnala
> The authors claims that the assembly, while made in pure python, is "fast enough" ? This shall be made more explicit. Do the authors mean that the assembly takes a negligible time compared to the resolution of the resulting linear system ?

There is now also a simple performance benchmark in the beginning of the README. Do you find it sufficient?
