### 2020-07-28 kinnala/scikit-fem#426 @kinnala
> Following the previous remark, it is not clear how the library could be used in this way. For example, is the mesh connectivity clearly described (how the unknowns are numbered) ?

We added documentation on the numbering of the DOF's at https://scikit-fem.readthedocs.io/en/latest/bcs.html
