### 2020-07-28 #2369 @kinnala
I have implemented a majority of the requested changes to the project README and other documentation; with the most significant effort being the introduction of a gallery of examples (https://scikit-fem.readthedocs.io/en/latest/listofexamples.html).

Next I'll focus on improving the paper itself which consist mainly of adding references and extending the discussion on the limits of the code's features. I hope to finish my work on the improvements by the end of the next week at the latest.
