### 2020-07-28 #2369 @meg-simula
Ok, thanks for the update!

On Tue, 28 Jul 2020 at 16:32, Tom Gustafsson <notifications@github.com>
wrote:

> I have implemented a majority of the requested changes to the project
> README and other documentation; with the most significant effort being the
> introduction of a gallery of examples (
> https://scikit-fem.readthedocs.io/en/latest/listofexamples.html).
>
> Next I'll focus on improving the paper itself which consist mainly of
> adding references and extending the discussion on the limits of the code's
> features. I hope to finish my work on the improvements by the end of the
> next week at the latest.
>
