### 2020-07-29 0c1ee6d
Add reference and other updates
