+++ paper/paper.bib
@@ -6,6 +6,16 @@
   doi = {10.5281/zenodo.1173115}
 }
 
+@book{fenicsbook,
+  title = {Automated Solution of Differential Equations by the Finite Element Method},
+  author = {Logg, Anders and Mardal, Kent-Andre and Wells, Garth N. and others},
+  editor = {Logg, Anders and Mardal, Kent-Andre and Wells, Garth N.},
+  year = {2012},
+  publisher = {Springer},
+  doi = {10.1007/978-3-642-23099-8},
+  isbn = {978-3-642-23098-1},
+}
+
 @article{fenics,
   title = {{The FEniCS Project version 1.5}},
   author = {Martin S. Aln{\ae}s and Jan Blechta and Johan Hake and August Johansson and Benjamin Kehlet and Anders Logg and Chris Richardson and Johannes Ring and Marie E. Rognes and Garth N. Wells},
