### 2020-07-29 bc1c8d1
Clarify software packages used in the creation of the figures
