+++ paper/paper.bib
@@ -165,3 +165,28 @@
   year = 2020,
   doi = {10.5281/zenodo.3862391},
 }
+
+@incollection{ahrens2005,
+  title = "36 - ParaView: An End-User Tool for Large-Data Visualization",
+  editor = "Charles D. Hansen and Chris R. Johnson",
+  booktitle = "Visualization Handbook",
+  publisher = "Butterworth-Heinemann",
+  address = "Burlington",
+  pages = "717--731",
+  year = "2005",
+  isbn = "978-0-12-387582-2",
+  doi = {10.1016/B978-012387582-2/50038-1},
+  author = "James Ahrens and Berk Geveci and Charles Law"
+}
+
+@article{hunter2007,
+  author = {Hunter, J. D.},
+  title = {Matplotlib: A 2D graphics environment},
+  journal = {Computing in Science \& Engineering},
+  volume = {9},
+  number = {3},
+  pages = {90--95},
+  publisher = {IEEE COMPUTER SOC},
+  doi = {10.1109/MCSE.2007.55},
+  year = 2007
+}
