+++ paper/paper.bib
@@ -167,9 +167,9 @@
 }
 
 @incollection{ahrens2005,
-  title = "36 - ParaView: An End-User Tool for Large-Data Visualization",
+  title = {{36 - ParaView: An End-User Tool for Large-Data Visualization}},
   editor = "Charles D. Hansen and Chris R. Johnson",
-  booktitle = "Visualization Handbook",
+  booktitle = {{Visualization Handbook}},
   publisher = "Butterworth-Heinemann",
   address = "Burlington",
   pages = "717--731",
