### 2020-07-29 6391a87
Add explicit GPLv3 license to the files depending on pygmsh
