+++ docs/examples/ex35.py
@@ -221,6 +221,24 @@ potential difference across the capacitor. Thus
 .. math::
    C = \\frac{2 E}{V^2}.
 
+License
+-------
+
+Copyright 2018-2020 scikit-fem developers
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 """
 from skfem.mesh import MeshTri
 from skfem.assembly import InteriorBasis, FacetBasis
