+++ docs/examples/ex12.py
@@ -23,6 +23,24 @@ problem, the maximum of the solution (normalized by the area) is the
 0.07363) and circles (k' = 1/Pi/4) at the centre and so can be
 evaluated by interpolation.
 
+License
+-------
+
+Copyright 2018-2020 scikit-fem developers
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 """
 
 from skfem import *
