### 2020-07-29 b248249
Add license exception
