+++ LICENSE.md
@@ -53,3 +53,25 @@ exceptions.
 > ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 > CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 > THE SOFTWARE.
+
+- docs/examples/ex{04,12,13,17,20,24,26,27,28,32,35}.py
+
+> Copyright 2018-2020 scikit-fem developers
+> 
+> This program is free software: you can redistribute it and/or modify
+> it under the terms of the GNU General Public License as published by
+> the Free Software Foundation, either version 3 of the License, or
+> (at your option) any later version.
+> 
+> This program is distributed in the hope that it will be useful,
+> but WITHOUT ANY WARRANTY; without even the implied warranty of
+> MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+> GNU General Public License for more details.
+> 
+> You should have received a copy of the GNU General Public License
+> along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+The above examples have a GPLv3 -licensed dependency.  These examples are not
+included in the source code distribution obtained via PyPI which will always
+remain free of GPLv3.
+
