### 2020-07-29 b670ff8
Minor improvements to the docs
