### 2020-07-30 35ad0ce
cholesky->splu, fixes #447
