### 2020-07-30 4bb5cfc
removed redundant note about needing scikit-sparse
