+++ docs/examples/ex10_nk.py
@@ -1,6 +1,18 @@
 """Nonlinear Poisson equation.
 
-This example solves the nonlinear minimal surface problem using a Newton-Krylov method.
+This example solves the nonlinear minimal surface problem using a preconditioned Jacobian-free Newton-Krylov method.
+
+The idea is (Knoll & Keyes 2004, &sect;5.1)
+
+    JFNK performs the Newton–Krylov iteration matrix-free, while the Jacobian used in the preconditioner is formed with a frequency of p Newton iterations.
+
+This is to be contrasted with (ibid.)
+
+    MNK (modified Newton–Krylov) is a standard Newton–Krylov method but the Jacobian is formed with a frequency of p Newton iterations for use in both the matrix vector multiply and the preconditioner.
+
+The advantage of the present method is that the finite-difference Fréchet approximation of the Jacobian in the Newton iteration is always accuate and up to date, only the quality of the preconditioner degrades with lagging.
+
+The Jacobian-free Newton-Krylov iteration is implemented in `scipy.optimize.root(method="krylov")` and an incomplete LU factorization of the lagged Jacobian is used as preconditioner.  The Jacobian is only reassembled and refactorized every p Newton iterations, and here by default p is infinite, in which case only the Jacobian based on the initial guess is ever used.
 
 """
 
@@ -9,6 +21,8 @@ from skfem.helpers import grad, dot
 from skfem.models.poisson import laplace
 import numpy as np
 from scipy.optimize import root
+from scipy.sparse.linalg import LinearOperator, spilu
+from typing import List, Optional
 
 m = MeshTri()
 m.refine(5)
@@ -17,7 +31,7 @@ m.refine(5)
 @BilinearForm
 def jacobian(u, v, w):
     return (1 / np.sqrt(1 + dot(grad(w['w']), grad(w['w']))) * dot(grad(u), grad(v))
-            -2 * dot(grad(u), grad(w['w'])) * dot(grad(w['w']), grad(v))
+            - 2 * dot(grad(u), grad(w['w'])) * dot(grad(w['w']), grad(v))
             / 2 / (1 + dot(grad(w['w']), grad(w['w'])))**(3/2))
 
 
@@ -32,22 +46,50 @@ x = np.zeros(basis.N)
 
 I = m.interior_nodes()
 D = m.boundary_nodes()
-x[D] = np.sin(np.pi * m.p[0, D]) 
+x[D] = np.sin(np.pi * m.p[0, D])
+
 
 def residual(u: np.ndarray) -> np.ndarray:
-    res = asm(rhs, basis, w=basis.interpolate(u))
-    res[D] = 0.
-    return res
+    u_full = np.empty_like(x)
+    u_full[I] = u
+    u_full[D] = x[D]
 
-M = build_pc_ilu(asm(jacobian, basis, w=basis.interpolate(x)))
+    return asm(rhs, basis, w=basis.interpolate(u_full))[I]
 
-sol = root(residual, x, method='krylov',
+
+M = LinearOperator([len(I)]*2, matvec=lambda u: u)
+
+
+def update(u: np.ndarray,
+           _: Optional[np.ndarray] = None,
+           i: List[int] = [0],
+           period: int = np.inf) -> None:
+
+    if i[0] % period == 0:
+
+        print('Updating Jacobian preconditioner.')
+
+        u_full = np.empty_like(x)
+        u_full[I] = u
+        u_full[D] = x[D]
+
+        J = asm(jacobian, basis, w=basis.interpolate(u_full))
+        JI = condense(J, D=D, expand=False).tocsc()
+        JI_ilu = spilu(JI)
+        M.matvec = JI_ilu.solve
+
+    i[0] += 1
+
+
[…]
