### 2020-07-30 8d6446d
Remove and reimplement code from meshzoo
