### 2020-07-30 b1051b6
Add licensing information to the README
