+++ paper/paper.md
@@ -49,19 +49,19 @@ various packages in the Python ecosystem such as meshio [@meshio], pacopy
 [@pacopy], and pyamg [@pyamg].
 
 In contrast to NGSolve [@ngsolve], FEniCS [@fenics], Firedrake [@firedrake],
-SfePy [@sfepy], and GetFEM [@getfem], `scikit-fem` adds no compiled code making
-the installation quick and straightforward.  We specifically target finite
-element assembly instead of encapsulating the entire finite element analysis
-from pre- to postprocessing into a single framework.  As a consequence, we
-cannot provide an end-to-end experience when it comes to, e.g., specific
+SfePy [@sfepy], and GetFEM [@getfem], `scikit-fem` contains no compiled code
+making the installation quick and straightforward.  We specifically target
+finite element assembly instead of encapsulating the entire finite element
+analysis from pre- to postprocessing into a single framework.  As a consequence,
+we cannot provide an end-to-end experience when it comes to, e.g., specific
 physical models or distributed computing.  Our aim is to be generic in terms of
 PDEs and, hence, support a variety of finite element schemes.  Currently
 `scikit-fem` includes basic support for $H^1$-, $H(\mathrm{div})$-,
 $H(\mathrm{curl})$-, and $H^2$-conforming problems as well as various
 nonconforming schemes.
 
-We support forms that depend on the values and the derivatives of the trial and
-the test functions, high-order derivatives, local mesh parameter, nonuniform
+We support weak forms that depend on the values and the derivatives of the trial
+and the test functions, high-order derivatives, local mesh parameter, nonuniform
 material fields, and existing finite element solutions.  Iterations related to,
 e.g., nonlinear problems (Newton's method and the variants, parameter
 continuation) or adaptive mesh refinement (evaluation of functionals, the
@@ -82,7 +82,7 @@ results from `scikit-fem`, e.g., @mcbain2018, @gustafsson2019, and
 @gustafsson2020.  In addition, `scikit-fem` is used in a recently published
 Python package for battery modelling [@pybamm].
 
-![(Top left.) A combination of triangular and quadrilateral elements is used to solve the linear elastic contact problem. The figure was created with the help of matplotlib [@hunter2007]. (Top right.) The lowest order tetrahedral Nédélec element is used to solve the $H(\mathrm{curl})$-conforming model problem $\nabla \times \nabla \times E + E = f$. The figure was created with the help of ParaView [@ahrens2005]. (Bottom.) The Taylor–Hood element is used to solve the Navier–Stokes flow over a backward-facing step for different Reynolds numbers. The figure was created with the help of matplotlib [@hunter2007].\label{fig:examples}](examples.png)
+![(Top left.) A combination of triangular and quadrilateral elements is used to solve the linear elastic contact problem. (Top right.) The lowest order tetrahedral Nédélec element is used to solve the $H(\mathrm{curl})$-conforming model problem $\nabla \times \nabla \times E + E = f$. The color corresponds to the magnitude of the vector field $E$. (Bottom.) The Taylor–Hood element is used to solve the Navier–Stokes flow over a backward-facing step for different Reynolds numbers.  The first and last two-dimensional figures were generated using scikit-fem's wrapping of matplotlib [@hunter2007]; three-dimensional postprocessing is more specialized and usually left to export of, e.g., VTK or XDMF, via meshio [@meshio] for subsequent rendering in, e.g., ParaView [@ahrens2005] as done in the second figure.\label{fig:examples}](examples.png)
 
 # Acknowledgements
 
