+++ paper/paper.md
@@ -82,7 +82,7 @@ results from `scikit-fem`, e.g., @mcbain2018, @gustafsson2019, and
 @gustafsson2020.  In addition, `scikit-fem` is used in a recently published
 Python package for battery modelling [@pybamm].
 
-![(Top left.) A combination of triangular and quadrilateral elements is used to solve the linear elastic contact problem. (Top right.) The lowest order tetrahedral Nédélec element is used to solve the $H(\mathrm{curl})$-conforming model problem $\nabla \times \nabla \times E + E = f$. The color corresponds to the magnitude of the vector field $E$. (Bottom.) The Taylor–Hood element is used to solve the Navier–Stokes flow over a backward-facing step for different Reynolds numbers.  The first and last two-dimensional figures were generated using scikit-fem's wrapping of matplotlib [@hunter2007]; three-dimensional postprocessing is more specialized and usually left to export of, e.g., VTK or XDMF, via meshio [@meshio] for subsequent rendering in, e.g., ParaView [@ahrens2005] as done in the second figure.\label{fig:examples}](examples.png)
+![(Top left.) A combination of triangular and quadrilateral elements is used to solve the linear elastic contact problem. (Top right.) The lowest order tetrahedral Nédélec element is used to solve the $H(\mathrm{curl})$-conforming model problem $\nabla \times \nabla \times E + E = f$. The color corresponds to the magnitude of the vector field $E$. (Bottom.) The Taylor–Hood element is used to solve the Navier–Stokes flow over a backward-facing step for different Reynolds numbers.  The first and last two-dimensional figures were generated using `scikit-fem`'s wrapping of matplotlib [@hunter2007]; three-dimensional postprocessing is more specialized and usually left to export of, e.g., VTK or XDMF formats, via meshio [@meshio] for subsequent rendering in, e.g., ParaView [@ahrens2005] as done in the second figure.\label{fig:examples}](examples.png)
 
 # Acknowledgements
 
