### 2020-08-01 4c2bc5e
Mention parallel processes and quadrature points
