### 2020-08-01 #2369 @kinnala
We thank both reviewers for their valuable feedback.  The submission has been
now updated based on the feedback.  Below are specific answers to the comments;
hopefully we were able to address everything.

# Answers to reviewer 1

> For someone who does not have a background in finite elements, they may not
> understand what basis functions or weak formulation means.

We added a reference to the FEniCS book which explains these concepts very well.

> The performance claim in vague:
> “Despite being fully interpreted, the code has a reasonably good performance.”
> If the author could be more specific (quantifiable), it would help the
> reviewer in making a judgement.

We added a benchmark to the README just below the given quote. The benchmark
compares time spent on assembly and on linear solve.

# Answers to reviewer 2

In the paper we added a paragraph discussing the various points mentioned by
@thelfer (nonuniform coefficients, boundary conditions, nonlinear problems,
fields at quadrature points, parallel computing, and so on).

> The library do provide a way to handle Dirichlet boundary conditions by
> elimination. This shall be at least mentioned

Done.

> For example, is the mesh connectivity clearly described (how the unknowns are
> numbered)?

Mesh connectivity is now covered in the package documentation:
https://scikit-fem.readthedocs.io/en/latest/bcs.html

> The authors shall describe how the library could be used in non linear cases.

It is now mentioned in the paper that nonlinear iterations should be implemented
by the user and that we provide tools and examples for doing that.

> Can non uniform coefficients be handled ?

Yes. It is now mentioned also in the paper.

> The authors claims that the assembly, while made in pure python, is "fast
> enough" ? This shall be made more explicit. Do the authors mean that the
> assembly takes a negligible time compared to the resolution of the resulting
> linear system ?

There is now a performance benchmark in the beginning of the README which
hopefully clarifies this.

> The authors claims that more than 30 examples illustrate the use of the
> library. I did not find any page that summarises those examples.

We added a page to the documentation which summarizes the examples:
https://scikit-fem.readthedocs.io/en/latest/listofexamples.html

> With which package have the figures been generated ?

Modified the caption to include this information.

> The description of the second example is very sparse.

Added the governing equation and clarified the meaning of the figure colors in
the caption.

> Some examples depends on external packages which are not installed with
> scikit-fem. An appropriate list of dependencies would be appreciated.

README has now has instructions on installing the test dependencies that allow
running all examples.

> UFL offers functional spaces at quadrature points. This functional spaces are
> relevant when dealing with non linear constitutive equations in solid
> mechanics. Is there anything similar in scikit-fem ?

We support assembling weak forms with arbitrary fields/values at the quadrature
points.  This is now mentioned in the paper also.

> Would it work in MPI-like processes ? Do you plan any extension of the package
> in that matter or is it out of its scope ?

Paper now clarifies our thinking on the distributed copmuting.  While it's out
of scope for `scikit-fem`, we can see it used as a building block in some kind
of specific parallel computations.
