### 2020-08-03 6b071d6
use global solution array
