### 2020-08-03 de32a2b
global solution array
