### 2020-08-05 kinnala/scikit-fem#426 @thelfer
> There is now also a simple performance benchmark in the beginning of the README. Do you find it sufficient?

That's fine for me. Thanks
