### 2020-08-05 kinnala/scikit-fem#426 @thelfer
I do appreciate our answers to my remarks and the time you spend on it. Closing this issue which is no more meaningful and start finalizing the review.
