### 2020-08-11 1ed0605
pin at pacopy 0.1.2; fixes #451
