+++ docs/listofexamples.rst
@@ -320,7 +320,7 @@ This example solves the Bratu-Gelfand two-point boundary value problem :math:`u'
 with :math:`u(0)=u(1)=0` and where :math:`\lambda > 0` is a parameter.
 
 .. note::
-   This example requires the external package `pacopy <https://github.com/nschloe/pacopy>`__.
+   This example requires the external package `pacopy 0.1.2 <https://pypi.org/project/pacopy/0.1.2>`__.
 
 .. figure:: https://user-images.githubusercontent.com/973268/87779278-38ccf700-c835-11ea-955a-b77a0336b791.png
 
@@ -375,13 +375,13 @@ See the `source code of Example 26 <https://github.com/kinnala/scikit-fem/blob/m
 Example 27: Backward-facing step
 ================================
 
-This example uses `pacopy <https://pypi.org/project/pacopy/>`__ to extend
+This example uses `pacopy 0.1.2 <https://pypi.org/project/pacopy/0.1.2>`__ to extend
 the Stokes equations over a backward-facing step (Example 24) to finite Reynolds
 number; this means defining a residual for the nonlinear problem and its
 derivatives with respect to the solution and to the Reynolds number.
 
 .. note::
-   This example requires the external packages `pygmsh <https://pypi.org/project/pygmsh/>`___ and `pacopy <https://pypi.org/project/pacopy/>`__.
+   This example requires the external packages `pygmsh <https://pypi.org/project/pygmsh/>`___ and `pacopy 0.1.2 <https://pypi.org/project/pacopy/0.1.2>`__.
 
 .. figure:: https://user-images.githubusercontent.com/973268/87858972-97c86400-c93a-11ea-86e4-66f870b03e48.png
 
