### 2020-08-11 b581561
remove pygmsh from ex27
