### 2020-08-11 fb3372d
remove special licence from ex27
