+++ docs/examples/ex27.py
@@ -1,10 +1,10 @@
 r"""Backward-facing step.
 
 .. warning::
-   This example requires the external packages `pygmsh <https://pypi.org/project/pygmsh/>`_ and `pacopy 0.1.2 <https://pypi.org/project/pacopy/0.1.2/>`_.
+   This example requires the external package `pygmsh <https://pypi.org/project/pygmsh/>`_.
 
 
-In this example, `pacopy 0.1.2 <https://pypi.org/project/pacopy/0.1.2/>`_ is used to extend
+In this example, natural parameter continuation is used to extend
 the Stokes equations over a backward-facing step to finite Reynolds
 number; this means defining a residual for the nonlinear problem and its
 derivatives with respect to the solution and to the parameter (here Reynolds
@@ -81,7 +81,181 @@ from scipy.sparse import bmat, block_diag, csr_matrix
 from pygmsh import generate_mesh
 from pygmsh.built_in import Geometry
 
-from pacopy import natural
+class NewtonConvergenceError(Exception):
+    """
+
+    Copied from pacopy 0.1.2, copyright N. Schlömer & G. D. McBain 2020 under
+    the MIT Licence.
+
+    https://raw.githubusercontent.com/nschloe/pacopy/v0.1.2/pacopy/natural.py
+
+    """
+    pass
+
+
+def newton(f, jacobian_solver, norm2, u0, tol=1.0e-10, max_iter=20, verbose=True):
+    """return a root of f near u0
+    
+    Copied from pacopy 0.1.2, copyright N. Schlömer & G. D. McBain 2020 under
+    the MIT Licence.
+
+    https://raw.githubusercontent.com/nschloe/pacopy/v0.1.2/pacopy/natural.py
+
+    """
+    
+    u = u0
+
+    fu = f(u)
+    nrm = np.sqrt(norm2(fu))
+    if verbose:
+        print(f"||F(u)|| = {nrm:e}")
+
+    k = 0
+    while k < max_iter:
+        if nrm < tol:
+            break
+        du = jacobian_solver(u, -fu)
+        u += du
+        fu = f(u)
+        nrm = np.sqrt(norm2(fu))
+        k += 1
+        if verbose:
+            print(f"||F(u)|| = {nrm:e}")
+
+    is_converged = nrm < tol
+
+    if not is_converged:
+        raise NewtonConvergenceError(
+            f"Newton's method didn't converge after {k} steps."
+        )
+
+    return u, k
+
+
+
+def natural(
+    problem,
+    u0,
+    lambda0,
+    callback,
+    lambda_stepsize0=1.0e-1,
+    lambda_stepsize_max=float("inf"),
+    lambda_stepsize_aggressiveness=2,
+    max_newton_steps=5,
+    newton_tol=1.0e-12,
+    max_steps=float("inf"),
+    verbose=True,
+    use_first_order_predictor=True,
+    milestones=None,
+):
+    """Natural parameter continuation.
+
+    The most naive parameter continuation. This simply solves :math:`F(u, \\lambda_0)=0`
+    using Newton's method, then changes :math:`\\lambda` slightly and solves again using
+    the previous solution as an initial guess. Cannot handle turning points.
+
+    Copied from pacopy 0.1.2, copyright N. Schlömer & G. D. McBain 2020 under
+    the MIT Licence.
+
+    https://raw.githubusercontent.com/nschloe/pacopy/v0.1.2/pacopy/natural.py
+
+    Args:
+        problem: Instance of the problem class
+        u0: Initial guess
+        lambda0: Initial parameter value
+        callback: Callback function
+        lambda_stepsize0 (float): Initial step size
+        lambda_stepsize_aggressiveness (float): The step size is adapted after each step
[…]
