### 2020-08-11 809b5e6
DOC: this example no longer depends on pygmsh
