### 2020-08-11 5adb748
DOC: ex27 no longer depends on pygmsh
