+++ skfem/assembly/__init__.py
@@ -55,8 +55,7 @@ from scipy.sparse import csr_matrix
 
 from .basis import Basis, InteriorBasis, FacetBasis
 from .dofs import Dofs, DofsView
-from .form import Form, BilinearForm, LinearForm, Functional,\
-    bilinear_form, linear_form, functional
+from .form import Form, BilinearForm, LinearForm, Functional
 
 
 def asm(form: Form,
@@ -78,7 +77,4 @@ __all__ = [
     "DofsView",
     "BilinearForm",
     "LinearForm",
-    "Functional",
-    "bilinear_form",
-    "linear_form",
-    "functional"]
+    "Functional"]
