+++ skfem/assembly/form/bilinear_form.py
@@ -1,4 +1,4 @@
-from typing import Callable, Optional, Any
+from typing import Optional, Any
 
 import numpy as np
 
