### 2020-08-18 074c9bb
share static mesh in exx24, 27
