### 2020-08-18 eb04922
exx24, 27 no longer need special licence
