### 2020-08-18 1ceb8ac
share static disk mesh amonst exx12, 17, 20 #458
