### 2020-08-18 2dbbd61
generalize disk.json subdomains & boundary names
