### 2020-08-18 ae5d059
export radii from ex17
