### 2020-08-18 94ae481
rewrite natural as generator
