+++ docs/examples/ex27.py
@@ -1,9 +1,5 @@
 r"""Backward-facing step.
 
-.. warning::
-   This example requires the external package `pacopy 0.1.2 <https://pypi.org/project/pacopy/0.1.2/>`_.
-
-
 In this example, natural parameter continuation is used to extend
 the Stokes equations over a backward-facing step to finite Reynolds
 number; this means defining a residual for the nonlinear problem and its
@@ -49,11 +45,11 @@ from skfem import *
 from skfem.helpers import grad, dot
 from skfem.models.poisson import vector_laplace, laplace
 from skfem.models.general import divergence, rot
-import skfem.io.json
+from skfem.io.json import from_file
 
 from functools import partial
 from pathlib import Path
-from typing import Tuple, Iterable
+from typing import Callable, Tuple, Iterable
 
 from matplotlib.pyplot import subplots
 from matplotlib.tri import Triangulation
@@ -74,65 +70,25 @@ class NewtonConvergenceError(Exception):
     pass
 
 
-# def newton(f, jacobian_solver, norm2, u0, tol=1.0e-10, max_iter=20, verbose=True):
-#     """return a root of f near u0
-    
-#     Copied from pacopy 0.1.2, copyright N. Schlömer & G. D. McBain 2020 under
-#     the MIT Licence.
-
-#     https://raw.githubusercontent.com/nschloe/pacopy/v0.1.2/pacopy/natural.py
-
-#     """
-    
-#     u = u0
-
-#     fu = f(u)
-#     nrm = np.sqrt(norm2(fu))
-#     if verbose:
-#         print(f"||F(u)|| = {nrm:e}")
-
-#     k = 0
-#     while k < max_iter:
-#         if nrm < tol:
-#             break
-#         du = jacobian_solver(u, -fu)
-#         u += du
-#         fu = f(u)
-#         nrm = np.sqrt(norm2(fu))
-#         k += 1
-#         if verbose:
-#             print(f"||F(u)|| = {nrm:e}")
-
-#     is_converged = nrm < tol
-
-#     if not is_converged:
-#         raise NewtonConvergenceError(
-#             f"Newton's method didn't converge after {k} steps."
-#         )
-
-#     return u, k
-
-
-
 def natural(
-    problem,
-    u0,
-    lambda0,
-    callback,
-    lambda_stepsize0=1.0e-1,
-    lambda_stepsize_max=float("inf"),
-    lambda_stepsize_aggressiveness=2,
-    max_newton_steps=5,
-    newton_tol=1e-9,
-    max_steps=float("inf"),
-    verbose=True,
-    use_first_order_predictor=True,
-    milestones=None,
-):
+    solver: Callable[[np.ndarray, float, float, int], OptimizeResult],
+    jacobian_solver: Callable[[np.ndarray, float, np.ndarray], np.ndarray],
+    df_dmu: Callable[[np.ndarray, float], np.ndarray],
+    u0: np.ndarray,
+    milestones: Iterable[float],
+    mu_stepsize0: float = 1.0e-1,
+    mu_stepsize_max: float = float("inf"),
+    mu_stepsize_aggressiveness: int = 2,
+    max_newton_steps: int = 5,
+    newton_tol: float = 1e-9,
+    max_steps: float = float("inf"),
+    verbose: bool = True,
+    use_first_order_predictor: bool = True
+) -> Iterable[Tuple[float, np.ndarray]]:
     """Natural parameter continuation.
 
-    The most naive parameter continuation. This simply solves :math:`F(u, \\lambda_0)=0`
-    using Newton's method, then changes :math:`\\lambda` slightly and solves again using
+    The most naive parameter continuation. This simply solves :math:`F(u, \\mu_0)=0`
[…]
