### 2020-08-18 e1aa319
no longer need pacopy for ex27
