+++ docs/listofexamples.rst
@@ -355,13 +355,10 @@ See the `source code of Example 26 <https://github.com/kinnala/scikit-fem/blob/m
 Example 27: Backward-facing step
 ================================
 
-This example uses `pacopy 0.1.2 <https://pypi.org/project/pacopy/0.1.2>`__ to extend
+This example uses natural parameter continuation to extend
 the Stokes equations over a backward-facing step (Example 24) to finite Reynolds
 number; this means defining a residual for the nonlinear problem and its
 derivatives with respect to the solution and to the Reynolds number.
-
-.. note::
-   This example requires the external package `pacopy 0.1.2 <https://pypi.org/project/pacopy/0.1.2>`__.
 
 .. figure:: https://user-images.githubusercontent.com/973268/87858972-97c86400-c93a-11ea-86e4-66f870b03e48.png
 
