### 2020-08-19 366c2ea
update use of project for Dirichlet data in ex27
