### 2020-08-19 75b7846
L2_projection->project, vectorially; fixes #460
