### 2020-08-20 #2369 @meg-simula
@kinnala This looks very nice, well done.

- [ ] Two very minor grammatical comments:  please replace (i) "FEM" by "the FEM" in the second paragraph and (ii) "examples on" by "examples by" at the top of page 2.

Next, could you please:
