### 2020-08-21 d8abd43
Bump up version number and make minor changes to the paper
