### 2020-08-21 #2369 @kinnala
Thanks @meg-simula . Hopefully I did the fixes correctly.

The tagged release happens to be 2.0.0. The major version increment does not signify other than that we happened to remove some (undocumented) features deprecated already 6 months ago while waiting for the review to finish, and hence semantic versioning mandates such an increment.

Here is the Zenodo archive: https://zenodo.org/record/3994106#.Xz-AhHX7Tb1

The corresponding DOI is: 10.5281/zenodo.3994106
