+++ docs/examples/ex36.py
@@ -1,0 +1,210 @@
+r"""Incompressible Hyperelasticity
+This example solves the governing equations describing the mechanical 
+response of a nearly incompressible elastomer using a mixed formulation. 
+The elastomer, assumed to be made up of a Neo-Hookean solid, occupies the domain 
+:math:`\Omega` in the undeformed configuration, with the internal stored energy 
+given by 
+.. math::
+   \int_\Omega\Psi(\mathbf{F})d\mathbf{X};\qquad\Psi(\mathbf{F}) = \mu/2 (I_1 - 3) - \mu \ln(J) + \lambda/2(J-1)^2
+
+where :math:`I_1 = \mathbf{F}:\mathbf{F} = tr(\mathbf{F}^T\mathbf{F})` and 
+:math:`J = \text{det}(\mathbf{F})` and :math:`\mathbf{F} = \mathbf{I} + \nabla\mathbf{u}` is the deformation gradient tensor. 
+The standard variational formulation in the absence of body force and surface traction can be written as
+.. math::
+   \min_{u\in\mathcal{K}} \int_\Omega \Psi(\mathbf{F})d\mathbf{X}
+
+where :math:`\mathcal{K}` is a set of kinematically admissible fields that satisfy the Dirichlet boundary condition(s). 
+However, this becomes ill-posed when :math:`\lambda/\mu\rightarrow +\infty`. In order to 
+circumvent this issue, we consider a mixed variational formulation, namely
+
+.. math::
+   \min_{\mathbf{u}\in\mathcal{K}}\max_{p}\int_\Omega \Psi*(\mathbf{F}, p)d\mathbf{X}
+
+where 
+
+.. math::
+   \Psi*(\mathbf{F}, p) = p (J-J*) + mu/2(I_1-3) - \mu\ln(J*) + \lambda/2(J*-1)^2
+
+and :math:`J* = (\lambda + p + \sqrt{(\lambda + p)^2 + 4\lambda\mu )}/(2\lambda)`. 
+The numerical solution to the above problem requires choosing stable finite element spaces for the displacement (:math:`\mathbf{u}`)
+and pressure (:math:`p`). The corresponding weak form is given by
+
+find :math:`(\mathbf{u},p)\in (V_1 x V_2)` such that 
+
+.. math::
+   \mathcal{F}_1 = \int_\Omega \left( \mu\mathbf{F} + p\mathbf{F}^{-T} \right) : \nabla\mathbf{v}d\mathbf{X}
+
+and 
+
+.. math::
+   \mathcal{F}_2 = \int_\Omega \frac{\partial\Psi*}{\partial p}q\ d\mathbf{X}
+
+for all :math:`(\mathbf{v},q)\in (V_1 x V_2)` and 
+
+.. math::
+   V_1 = \left\{\mathbf{u}\ni \mathbf{u}\in (H^1(\Omega))^3 \cap u\in\mathcal{K} \right\}
+
+and
+
+.. math::
+   V_2 = \left\{ p\ni p\in L^2(\Omega) \right\}
+
+Here, inspired by it's counterpart in fluid mechanics, we choose the lowest order Taylor-Hood element (:math:`\mathbb{P}_2-P_1`) 
+which satifies the Babuska-Brezzi condition. Fore more details on the derivation, see http://pamies.cee.illinois.edu/Publications_files/IJNME_2015.pdf#page=4&zoom=100,312,414
[…]
