### 2020-08-23 1a514ea
fix typos in description
