### 2020-08-24 d65be2e
fixed typo in the definition of F1 and F2
