+++ skfem/helpers.py
@@ -134,11 +134,9 @@ def inv(A):
         Inverse of `A`.
     """
     invA = zeros_like(A)
-    detA = A[0, 0] * (A[1, 1] * A[2, 2] -\
-        A[1, 2] * A[2, 1]) -\
-            A[0, 1] * (A[1, 0] * A[2, 2] - A[1, 2] * A[2, 0]) +\
-                A[0, 2] * (A[1, 0] * A[2, 1] -\
-                    A[1, 1] * A[2, 0])
+    detA = A[0, 0] * (A[1, 1] * A[2, 2] - A[1, 2] * A[2, 1]) -\
+        A[0, 1] * (A[1, 0] * A[2, 2] - A[1, 2] * A[2, 0]) +\
+            A[0, 2] * (A[1, 0] * A[2, 1] - A[1, 1] * A[2, 0])
 
 
     invA[0, 0] = (-A[1, 2] * A[2, 1] +
