+++ skfem/helpers.py
@@ -113,11 +113,11 @@ def det(A):
     """
     detA = zeros_like(A[0, 0])
     detA = A[0, 0] * (A[1, 1] * A[2, 2] -
-                    A[1, 2] * A[2, 1]) -\
-        A[0, 1] * (A[1, 0] * A[2, 2] -
-                    A[1, 2] * A[2, 0]) +\
-        A[0, 2] * (A[1, 0] * A[2, 1] -
-                    A[1, 1] * A[2, 0])
+                      A[1, 2] * A[2, 1]) -\
+           A[0, 1] * (A[1, 0] * A[2, 2] -
+                      A[1, 2] * A[2, 0]) +\
+           A[0, 2] * (A[1, 0] * A[2, 1] -
+                      A[1, 1] * A[2, 0])
     return detA
 
 
@@ -135,26 +135,25 @@ def inv(A):
     """
     invA = zeros_like(A)
     detA = A[0, 0] * (A[1, 1] * A[2, 2] - A[1, 2] * A[2, 1]) -\
-        A[0, 1] * (A[1, 0] * A[2, 2] - A[1, 2] * A[2, 0]) +\
+            A[0, 1] * (A[1, 0] * A[2, 2] - A[1, 2] * A[2, 0]) +\
             A[0, 2] * (A[1, 0] * A[2, 1] - A[1, 1] * A[2, 0])
 
-
     invA[0, 0] = (-A[1, 2] * A[2, 1] +
-                A[1, 1] * A[2, 2]) / detA
+                  A[1, 1] * A[2, 2]) / detA
     invA[1, 0] = (A[1, 2] * A[2, 0] -
-                A[1, 0] * A[2, 2]) / detA
+                  A[1, 0] * A[2, 2]) / detA
     invA[2, 0] = (-A[1, 1] * A[2, 0] +
-                A[1, 0] * A[2, 1]) / detA
+                  A[1, 0] * A[2, 1]) / detA
     invA[0, 1] = (A[0, 2] * A[2, 1] -
-                A[0, 1] * A[2, 2]) / detA
+                  A[0, 1] * A[2, 2]) / detA
     invA[1, 1] = (-A[0, 2] * A[2, 0] +
-                A[0, 0] * A[2, 2]) / detA
+                  A[0, 0] * A[2, 2]) / detA
     invA[2, 1] = (A[0, 1] * A[2, 0] -
-                A[0, 0] * A[2, 1]) / detA
+                  A[0, 0] * A[2, 1]) / detA
     invA[0, 2] = (-A[0, 2] * A[1, 1] +
-                A[0, 1] * A[1, 2]) / detA
+                  A[0, 1] * A[1, 2]) / detA
     invA[1, 2] = (A[0, 2] * A[1, 0] -
-                A[0, 0] * A[1, 2]) / detA
+                  A[0, 0] * A[1, 2]) / detA
     invA[2, 2] = (-A[0, 1] * A[1, 0] +
-                A[0, 0] * A[1, 1]) / detA
+                  A[0, 0] * A[1, 1]) / detA
     return invA
