+++ skfem/helpers.py
@@ -114,9 +114,9 @@ def det(A):
     detA = zeros_like(A[0, 0])
     detA = A[0, 0] * (A[1, 1] * A[2, 2] -
                       A[1, 2] * A[2, 1]) -\
-           A[0, 1] * (A[1, 0] * A[2, 2] -
+          A[0, 1] * (A[1, 0] * A[2, 2] -
                       A[1, 2] * A[2, 0]) +\
-           A[0, 2] * (A[1, 0] * A[2, 1] -
+          A[0, 2] * (A[1, 0] * A[2, 1] -
                       A[1, 1] * A[2, 0])
     return detA
 
@@ -135,8 +135,8 @@ def inv(A):
     """
     invA = zeros_like(A)
     detA = A[0, 0] * (A[1, 1] * A[2, 2] - A[1, 2] * A[2, 1]) -\
-            A[0, 1] * (A[1, 0] * A[2, 2] - A[1, 2] * A[2, 0]) +\
-            A[0, 2] * (A[1, 0] * A[2, 1] - A[1, 1] * A[2, 0])
+           A[0, 1] * (A[1, 0] * A[2, 2] - A[1, 2] * A[2, 0]) +\
+           A[0, 2] * (A[1, 0] * A[2, 1] - A[1, 1] * A[2, 0])
 
     invA[0, 0] = (-A[1, 2] * A[2, 1] +
                   A[1, 1] * A[2, 2]) / detA
