### 2020-08-25 38f0c10
modify helpers to include 2,2 matrices
