+++ skfem/helpers.py
@@ -116,12 +116,12 @@ def det(A):
     if A.shape[0] == 3:
         detA = A[0, 0] * (A[1, 1] * A[2, 2] -
                         A[1, 2] * A[2, 1]) -\
-            A[0, 1] * (A[1, 0] * A[2, 2] -
+               A[0, 1] * (A[1, 0] * A[2, 2] -
                     A[1, 2] * A[2, 0]) +\
-            A[0, 2] * (A[1, 0] * A[2, 1] -
+               A[0, 2] * (A[1, 0] * A[2, 1] -
                     A[1, 1] * A[2, 0])
     elif A.shape[0] == 2:
-        detA = A[0,0] * A[1,1] - A[1,0] * A[0,1]
+        detA = A[0, 0] * A[1, 1] - A[1, 0] * A[0, 1]
     return detA
 
 
