+++ skfem/helpers.py
@@ -115,11 +115,11 @@ def det(A):
     detA = zeros_like(A[0, 0])
     if A.shape[0] == 3:
         detA = A[0, 0] * (A[1, 1] * A[2, 2] -
-                        A[1, 2] * A[2, 1]) -\
+                          A[1, 2] * A[2, 1]) -\
                A[0, 1] * (A[1, 0] * A[2, 2] -
-                    A[1, 2] * A[2, 0]) +\
+                          A[1, 2] * A[2, 0]) +\
                A[0, 2] * (A[1, 0] * A[2, 1] -
-                    A[1, 1] * A[2, 0])
+                          A[1, 1] * A[2, 0])
     elif A.shape[0] == 2:
         detA = A[0, 0] * A[1, 1] - A[1, 0] * A[0, 1]
     return detA
@@ -142,23 +142,23 @@ def inv(A):
     detA = det(A)
     if A.shape[0] == 3:
         invA[0, 0] = (-A[1, 2] * A[2, 1] +
-                    A[1, 1] * A[2, 2]) / detA
+                      A[1, 1] * A[2, 2]) / detA
         invA[1, 0] = (A[1, 2] * A[2, 0] -
-                    A[1, 0] * A[2, 2]) / detA
+                      A[1, 0] * A[2, 2]) / detA
         invA[2, 0] = (-A[1, 1] * A[2, 0] +
-                    A[1, 0] * A[2, 1]) / detA
+                      A[1, 0] * A[2, 1]) / detA
         invA[0, 1] = (A[0, 2] * A[2, 1] -
-                    A[0, 1] * A[2, 2]) / detA
+                      A[0, 1] * A[2, 2]) / detA
         invA[1, 1] = (-A[0, 2] * A[2, 0] +
-                    A[0, 0] * A[2, 2]) / detA
+                      A[0, 0] * A[2, 2]) / detA
         invA[2, 1] = (A[0, 1] * A[2, 0] -
-                    A[0, 0] * A[2, 1]) / detA
+                      A[0, 0] * A[2, 1]) / detA
         invA[0, 2] = (-A[0, 2] * A[1, 1] +
-                    A[0, 1] * A[1, 2]) / detA
+                      A[0, 1] * A[1, 2]) / detA
         invA[1, 2] = (A[0, 2] * A[1, 0] -
-                    A[0, 0] * A[1, 2]) / detA
+                      A[0, 0] * A[1, 2]) / detA
         invA[2, 2] = (-A[0, 1] * A[1, 0] +
-                    A[0, 0] * A[1, 1]) / detA
+                      A[0, 0] * A[1, 1]) / detA
     elif A.shape[0] == 2:
         invA[0, 0] = A[1, 1] / detA
         invA[0, 1] = -A[0, 1] / detA
