+++ docs/examples/ex36.py
@@ -1,4 +1,5 @@
 r"""Incompressible Hyperelasticity
+
 This example solves the governing equations describing the mechanical 
 response of a nearly incompressible elastomer using a mixed formulation. 
 The elastomer, assumed to be made up of a Neo-Hookean solid, occupies the domain 
@@ -7,15 +8,17 @@ given by
 .. math::
    \int_\Omega\Psi(\mathbf{F})d\mathbf{X};\qquad\Psi(\mathbf{F}) = \mu/2 (I_1 - 3) - \mu \ln(J) + \lambda/2(J-1)^2
 
-where :math:`I_1 = \mathbf{F}:\mathbf{F} = tr(\mathbf{F}^T\mathbf{F})` and 
-:math:`J = \text{det}(\mathbf{F})` and :math:`\mathbf{F} = \mathbf{I} + \nabla\mathbf{u}` is the deformation gradient tensor. 
-The standard variational formulation in the absence of body force and surface traction can be written as
+where :math:`I_1 = \mathbf{F}:\mathbf{F} = tr(\mathbf{F}^T\mathbf{F})` and
+:math:`J = \text{det}(\mathbf{F})` and :math:`\mathbf{F} = \mathbf{I} +
+\nabla\mathbf{u}` is the deformation gradient tensor.  The standard variational
+formulation in the absence of body force and surface traction can be written as
 .. math::
    \min_{u\in\mathcal{K}} \int_\Omega \Psi(\mathbf{F})d\mathbf{X}
 
-where :math:`\mathcal{K}` is a set of kinematically admissible fields that satisfy the Dirichlet boundary condition(s). 
-However, this becomes ill-posed when :math:`\lambda/\mu\rightarrow +\infty`. In order to 
-circumvent this issue, we consider a mixed variational formulation, namely
+where :math:`\mathcal{K}` is a set of kinematically admissible fields that
+satisfy the Dirichlet boundary condition(s).  However, this becomes ill-posed
+when :math:`\lambda/\mu\rightarrow +\infty`. In order to circumvent this issue,
+we consider a mixed variational formulation, namely
 
 .. math::
    \min_{\mathbf{u}\in\mathcal{K}}\max_{p}\int_\Omega \Psi*(\mathbf{F}, p)d\mathbf{X}
@@ -25,9 +28,10 @@ where
 .. math::
    \Psi*(\mathbf{F}, p) = p (J-J*) + mu/2(I_1-3) - \mu\ln(J*) + \lambda/2(J*-1)^2
 
-and :math:`J* = (\lambda + p + \sqrt{(\lambda + p)^2 + 4\lambda\mu )}/(2\lambda)`. 
-The numerical solution to the above problem requires choosing stable finite element spaces for the displacement (:math:`\mathbf{u}`)
-and pressure (:math:`p`). The corresponding weak form is given by
+and :math:`J* = (\lambda + p + \sqrt{(\lambda + p)^2 + 4\lambda\mu)}/(2\lambda)`.
+The numerical solution to the above problem requires choosing stable finite
+element spaces for the displacement (:math:`\mathbf{u}`) and pressure
+(:math:`p`). The corresponding weak form is given by
 
 find :math:`(\mathbf{u},p)\in (V_1 x V_2)` such that 
 
@@ -49,46 +53,61 @@ and
 .. math::
    V_2 = \left\{ p\ni p\in L^2(\Omega) \right\}
 
[…]
