### 2020-08-31 fbdb564
Reduce the amount of decimals
