+++ README.md
@@ -29,21 +29,21 @@ of `scipy.sparse.linalg.spsolve`.*
 
 | Degrees-of-freedom | Time spent on assembly (s) | Time spent on linear solve (s) |
 | --- | --- | --- |
-| 64 | 0.001719430962111801 | 0.0008631939999759197 |
-| 125 | 0.0022377190180122852 | 0.0009978360030800104 |
-| 216 | 0.004175334994215518 | 0.0016233620117418468 |
-| 512 | 0.006831337988842279 | 0.0018000249983742833 |
-| 1000 | 0.027329800999723375 | 0.006487983977422118 |
-| 1728 | 0.02035172702744603 | 0.00894418201642111 |
-| 4096 | 0.04277183400699869 | 0.054088378965388983 |
-| 8000 | 0.08925403101602569 | 0.39822074701078236 |
-| 15625 | 0.17675437999423593 | 3.5216876140329987 |
-| 32768 | 0.41025503899436444 | 24.167141190962866 |
-| 64000 | 0.8814217780018225 | 138.6927448490169 |
-| 125000 | 1.7890089299762622 | nan |
-| 262144 | 4.044992835028097 | nan |
-| 512000 | 8.037139102991205 | nan |
-| 1030301 | 20.241967056994326 | nan |
+| 64 | 0.00155 | 0.00073 |
+| 125 | 0.00203 | 0.00072 |
+| 216 | 0.00276 | 0.00081 |
+| 512 | 0.00589 | 0.00127 |
+| 1000 | 0.01076 | 0.00247 |
+| 1728 | 0.02063 | 0.00538 |
+| 4096 | 0.04805 | 0.04241 |
+| 8000 | 0.09804 | 0.16269 |
+| 15625 | 0.20347 | 0.87741 |
+| 32768 | 0.46399 | 5.98163 |
+| 64000 | 1.00143 | 36.47855 |
+| 125000 | 2.05274 | nan |
+| 262144 | 4.48825 | nan |
+| 512000 | 8.82814 | nan |
+| 1030301 | 18.25461 | nan |
 
 
 ## Examples
