+++ README.md
@@ -1,6 +1,6 @@
 ![scikit-fem](https://github.com/kinnala/scikit-fem/blob/master/logo.png?raw=true)
 
-[![PyPI](https://img.shields.io/pypi/v/scikit-fem)](https://badge.fury.io/py/scikit-fem)
+[![PyPI](https://img.shields.io/pypi/v/scikit-fem)](https://pypi.org/project/scikit-fem/)
 [![Build Status](https://travis-ci.com/kinnala/scikit-fem.svg?branch=master)](https://travis-ci.com/kinnala/scikit-fem)
 [![License](https://img.shields.io/badge/license-BSD%203--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause)
 [![status](https://joss.theoj.org/papers/4120aba1525403e6d0972f4270d7b61e/status.svg)](https://joss.theoj.org/papers/4120aba1525403e6d0972f4270d7b61e)
