### 2020-09-01 e473e97
delete spurious second script
