### 2020-09-07 558190e
new continuation algorithm
