### 2020-09-07 a41b7e3
use new natural_jfnk in ex27
