### 2020-09-08 d5167f2
custom norm scipy/scipy#12814
