### 2020-09-08 823eba7
maxiter is O. K. just don't exceed it
