+++ skfem/algorithms/continuation.py
@@ -42,7 +42,7 @@ def natural_jfnk(
                 options["jac_options"].update(inner_M=preconditioner(mu, w))
 
             sol = root(fun(mu), sol.x, method="krylov", options=options)
-            if sol.nit >= options["maxiter"]:
+            if sol.nit > options["maxiter"]:
                 print(f"No convergence for mu={mu}.")
                 mu -= d_mu
                 d_mu /= 2
