### 2020-09-08 ec0ebb8
simplify root options
