### 2020-09-12 5cde703
Start implementing ElementHex2
