+++ skfem/element/element_hex/element_hex2.py
@@ -1,0 +1,131 @@
+import numpy as np
+from numpy.linalg import inv
+
+from ..element_h1 import ElementH1
+from ...mesh.mesh3d import MeshHex
+
+
+class ElementHexS2(ElementH1):
+    nodal_dofs = 1
+    facet_dofs = 1
+    edge_dofs = 1
+    interior_dofs = 1
+    dim = 3
+    maxdeg = 4
+    dofnames = ['u', 'u', 'u', 'u']
+    doflocs = np.array([[1., 1., 1.],
+                        [1., 1., 0.],
+                        [1., 0., 1.],
+                        [0., 1., 1.],
+                        [1., 0., 0.],
+                        [0., 1., 0.],
+                        [0., 0., 1.],
+                        [0., 0., 0.],
+                        [1., .5, .5],
+                        [.5, .5, 1.],
+                        [.5, 1., .5],
+                        [.5, .0, .5],
+                        [.5, .5, 0.],
+                        [0., .5, .5],
+                        [1., 1., .5],
+                        [1., .5, 1.],
+                        [.5, 1., 1.],
+                        [1., .5, 0.],
+                        [.5, 1., 0.],
+                        [1., 0., .5],
+                        [.5, 0., 1.],
+                        [0., 1., .5],
+                        [0., .5, 1.],
+                        [.5, 0., 0.],
+                        [0., .5, 0.],
+                        [0., 0., .5],
+                        [.5, .5, .5]])
+    mesh_type = MeshHex
+
+    def __init__(self):
+
+        X = self.doflocs.T
+        V = np.array([
+            1. + 0. * X[0],
+            X[0],
+            X[1],
+            X[2],
+            X[0] ** 2,
+            X[1] ** 2,
+            X[2] ** 2,
+            X[0] * X[1],
+            X[1] * X[2],
+            X[0] * X[2],
+            X[0] ** 2 * X[1],
+            X[1] ** 2 * X[2],
+            X[0] ** 2 * X[2],
+            X[0] * X[1] ** 2,
+            X[1] * X[2] ** 2,
+            X[0] * X[2] ** 2,
+            X[0] ** 2 * X[1] ** 2,
+            X[1] ** 2 * X[2] ** 2,
+            X[0] ** 2 * X[2] ** 2,
+            X[0] * X[1] * X[2],
+            X[0] ** 2 * X[1] * X[2],
+            X[0] * X[1] ** 2 * X[2],
+            X[0] * X[1] * X[2] ** 2,
+            X[0] ** 2 * X[1] ** 2 * X[2],
+            X[0] * X[1] ** 2 * X[2] ** 2,
+            X[0] ** 2 * X[1] * X[2] ** 2,
+            X[0] ** 2 * X[1] ** 2 * X[2] ** 2,
+        ])
+        self.invV = inv(V)
+
+    def lbasis(self, X, i):
+        x, y, z = 2 * X - 1
+
+        if i < 8:
+            s = [
+                (1, 1, 1),
+                (1, 1, -1),
+                (1, -1, 1),
+                (-1, 1, 1),
+                (1, -1, -1),
+                (-1, 1, -1),
+                (-1, -1, 1),
+                (-1, -1, -1),
+            ][i]
+            x *= s[0]
+            y *= s[1]
+            z *= s[2]
+            phi = (1 + x) * (1 + y) * (1 + z) * (x + y + z - 2) / 8
+            dphi = np.array([s[0] * (1 + y) * (1 + z) * (x + y + z - 2)
+                             + s[0] * (1 + x) * (1 + y) * (1 + z),
+                             s[1] * (1 + x) * (1 + z) * (x + y + z - 2)
[…]
