+++ skfem/element/element_hex/element_hex2.py
@@ -5,13 +5,14 @@ from ..element_h1 import ElementH1
 from ...mesh.mesh3d import MeshHex
 
 
-class ElementHexS2(ElementH1):
+class ElementHex2(ElementH1):
+
     nodal_dofs = 1
     facet_dofs = 1
     edge_dofs = 1
     interior_dofs = 1
     dim = 3
-    maxdeg = 4
+    maxdeg = 6
     dofnames = ['u', 'u', 'u', 'u']
     doflocs = np.array([[1., 1., 1.],
                         [1., 1., 0.],
@@ -43,9 +44,13 @@ class ElementHexS2(ElementH1):
     mesh_type = MeshHex
 
     def __init__(self):
-
         X = self.doflocs.T
-        V = np.array([
+        V = self._power_basis(X)
+        self.invV = inv(V).T
+        self.invV[np.abs(self.invV) < 1e-10] = 0.
+
+    def _power_basis(self, X):
+        return np.array([
             1. + 0. * X[0],
             X[0],
             X[1],
@@ -74,58 +79,106 @@ class ElementHexS2(ElementH1):
             X[0] ** 2 * X[1] * X[2] ** 2,
             X[0] ** 2 * X[1] ** 2 * X[2] ** 2,
         ])
-        self.invV = inv(V)
+
+    def _power_basis_dx(self, X):
+        return np.array([
+            0. * X[0],
+            1. + 0. * X[0],
+            0. * X[0],
+            0. * X[0],
+            2. * X[0],
+            0. * X[0],
+            0. * X[0],
+            X[1],
+            0. * X[0],
+            X[2],
+            2. * X[0] * X[1],
+            0. * X[0],
+            2. * X[0] * X[2],
+            X[1] ** 2,
+            0. * X[0],
+            X[2] ** 2,
+            2. * X[0] * X[1] ** 2,
+            0. * X[0],
+            2. * X[0] * X[2] ** 2,
+            X[1] * X[2],
+            2. * X[0] * X[1] * X[2],
+            X[1] ** 2 * X[2],
+            X[1] * X[2] ** 2,
+            2. * X[0] * X[1] ** 2 * X[2],
+            X[1] ** 2 * X[2] ** 2,
+            2. * X[0] * X[1] * X[2] ** 2,
+            2. * X[0] * X[1] ** 2 * X[2] ** 2,
+        ])
+
+    def _power_basis_dy(self, X):
+        return np.array([
+            0. * X[0],
+            0. * X[0],
+            1. + 0. * X[0],
+            0. * X[0],
+            0. * X[0],
+            2. * X[1],
+            0. * X[0],
+            X[0],
+            X[2],
+            0. * X[0],
+            X[0] ** 2,
+            2. * X[1] * X[2],
+            0. * X[0],
+            X[0] * 2. * X[1],
+            X[2] ** 2,
+            0. * X[0],
+            X[0] ** 2 * 2. * X[1],
+            2. * X[1] * X[2] ** 2,
+            0. * X[0],
+            X[0] * X[2],
+            X[0] ** 2 * X[2],
+            X[0] * 2. *  X[1] * X[2],
+            X[0] * X[2] ** 2,
+            X[0] ** 2 * 2. * X[1] * X[2],
+            X[0] * 2. * X[1] * X[2] ** 2,
+            X[0] ** 2 * X[2] ** 2,
+            X[0] ** 2 * 2. * X[1] * X[2] ** 2,
+        ])
+
+    def _power_basis_dz(self, X):
+        return np.array([
+            0. * X[0],
+            0. * X[0],
+            0. * X[0],
+            1. + 0. * X[0],
+            0. * X[0],
+            0. * X[0],
+            2. * X[2],
+            0. * X[0],
+            X[1],
+            X[0],
[…]
