+++ skfem/element/__init__.py
@@ -42,7 +42,7 @@ from .element_quad import ElementQuad0, ElementQuad1, ElementQuad2,\
     ElementQuadBFS
 from .element_tet import ElementTetP0, ElementTetP1, ElementTetP2,\
     ElementTetRT0, ElementTetN0, ElementTetMini
-from .element_hex import ElementHex1, ElementHexS2  # noqa
+from .element_hex import ElementHex1, ElementHex2, ElementHexS2  # noqa
 from .element_line import ElementLineP1, ElementLineP2,\
     ElementLinePp, ElementLineHermite, ElementLineMini  # noqa
 from .element_composite import ElementComposite  # noqa
@@ -77,6 +77,7 @@ __all__ = [
     "ElementTetN0",
     "ElementTetMini",
     "ElementHex1",
+    "ElementHex2",
     "ElementHexS2",
     "ElementLineP1",
     "ElementLineP2",
