### 2020-09-13 bb29a16
More tests and fixes to ElementHex2
