+++ skfem/element/element_tet/element_tet_n0.py
@@ -4,6 +4,7 @@ from ...mesh.mesh3d import MeshTet
 
 
 class ElementTetN0(ElementHcurl):
+
     edge_dofs = 1
     dim = 3
     maxdeg = 1
