### 2020-09-13 b444e02
Move ElementLinePp initialization test to test_elements
