### 2020-09-13 a5d2dcf
Clip inverse mapping Newton iteration to a range for robustness
