### 2020-09-13 cdf7384
Fix order of doflocs
