+++ skfem/element/element_hex/element_hex2.py
@@ -22,13 +22,7 @@ class ElementHex2(ElementH1):
                         [0., 1., 0.],
                         [0., 0., 1.],
                         [0., 0., 0.],
-                        [1., .5, .5],
-                        [.5, .5, 1.],
-                        [.5, 1., .5],
-                        [.5, .0, .5],
-                        [.5, .5, 0.],
-                        [0., .5, .5],
-                        [1., 1., .5],
+                        [1., 1., .5], # edges
                         [1., .5, 1.],
                         [.5, 1., 1.],
                         [1., .5, 0.],
@@ -40,6 +34,12 @@ class ElementHex2(ElementH1):
                         [.5, 0., 0.],
                         [0., .5, 0.],
                         [0., 0., .5],
+                        [1., .5, .5], # facets
+                        [.5, .5, 1.],
+                        [.5, 1., .5],
+                        [.5, .0, .5],
+                        [.5, .5, 0.],
+                        [0., .5, .5],
                         [.5, .5, .5]])
     mesh_type = MeshHex
 
