### 2020-09-13 dc59bd2
Fix flake8 issues
