+++ skfem/element/element_hex/element_hex2.py
@@ -22,7 +22,7 @@ class ElementHex2(ElementH1):
                         [0., 1., 0.],
                         [0., 0., 1.],
                         [0., 0., 0.],
-                        [1., 1., .5], # edges
+                        [1., 1., .5],  # edges
                         [1., .5, 1.],
                         [.5, 1., 1.],
                         [1., .5, 0.],
@@ -34,7 +34,7 @@ class ElementHex2(ElementH1):
                         [.5, 0., 0.],
                         [0., .5, 0.],
                         [0., 0., .5],
-                        [1., .5, .5], # facets
+                        [1., .5, .5],  # facets
                         [.5, .5, 1.],
                         [.5, 1., .5],
                         [.5, .0, .5],
@@ -134,7 +134,7 @@ class ElementHex2(ElementH1):
             0. * X[0],
             X[0] * X[2],
             X[0] ** 2 * X[2],
-            X[0] * 2. *  X[1] * X[2],
+            X[0] * 2. * X[1] * X[2],
             X[0] * X[2] ** 2,
             X[0] ** 2 * 2. * X[1] * X[2],
             X[0] * 2. * X[1] * X[2] ** 2,
