### 2020-09-13 996b0b2
Raise error if meshio mesh type not supported
