### 2020-09-14 f8924ef
Raise ValueError in ElementQuadBFS for wrong indices
