### 2020-09-15 8403c31
Minor edits to the docs
