### 2020-09-16 e292734
DOC: ex12 hasn't used an external mesher since #458
