### 2020-09-16 9eae8b9
use MeshTri.init_circle in exx; fixes #476
