### 2020-09-16 53010b4
update figures for examples newly using init_circle
