+++ docs/listofexamples.rst
@@ -146,7 +146,7 @@ Example 12: Postprocessing
 
 This example demonstrates postprocessing the value of a functional, Boussinesq's k-factor.
 
-.. figure:: https://user-images.githubusercontent.com/973268/87686059-bee13300-c78c-11ea-9693-727f0baf0433.png
+.. figure:: https://user-images.githubusercontent.com/1588947/93292071-0127fe80-f828-11ea-8c9e-46590d280b69.png
 
    The solution of Example 12.
 
@@ -235,7 +235,7 @@ This example solves for the creeping flow problem in the primitive variables,
 i.e. velocity and pressure instead of the stream-function.  These are governed
 by the Stokes momentum :math:`- \nu\Delta\boldsymbol{u} + \rho^{-1}\nabla p = \boldsymbol{f}` and the continuity equation :math:`\nabla\cdot\boldsymbol{u} = 0`.
 
-.. figure:: https://user-images.githubusercontent.com/973268/87775390-b04b5800-c82e-11ea-8999-e22305e909c1.png
+.. figure:: https://user-images.githubusercontent.com/1588947/93292002-d6d64100-f827-11ea-9a0a-c64d5d2979b7.png
 
    The streamlines of Example 18.
 
@@ -266,7 +266,7 @@ conditions at a wall are that :math:`\psi` is constant (the wall is
 impermeable) and that the normal component of its gradient vanishes (no
 slip)
 
-.. figure:: https://user-images.githubusercontent.com/973268/87778910-9745a580-c834-11ea-8277-62d58a7fe7b8.png
+.. figure:: https://user-images.githubusercontent.com/1588947/93291998-d50c7d80-f827-11ea-861b-f24ed27072d0.png
 
    The velocity field of Example 20.
 
