### 2020-09-17 fa45abf
Use new logo
