### 2020-09-17 47b9255
Reduce width of logo
