### 2020-09-17 7be12cb
Use image tag for resize
