### 2020-09-17 d3b942f
Make logo even smaller
