### 2020-09-17 e7ee94e
Add missing changelog enetry
