+++ README.md
@@ -170,6 +170,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 - `ElementHex2`, a triquadratic hexahedral element
+- `MeshTri.init_circle`, constructor for a circle mesh
 
 ### [2.0.0] - 2020-08-21
 
