### 2020-09-30 c82cc77
Pin pygmsh to 6.1.1 in the test suite
