+++ requirements.txt
@@ -5,6 +5,6 @@ pytest
 ipython
 sphinx
 meshio
-pygmsh
+pygmsh==6.1.1
 pacopy==0.1.2
 sphinx_rtd_theme
