### 2020-09-30 a8c390a
Optimize boundary_edges
