### 2020-09-30 45e3cf9
Fix the test and update changelog
