+++ skfem/mesh/mesh3d/mesh3d.py
@@ -38,7 +38,7 @@ class Mesh3D(Mesh):
                               self.facets[(itr + 1) % self.facets.shape[0],
                               facets]))
                    for itr in range(self.facets.shape[0])])).T, axis=1)
-        edge_candidates = self.t2e[:, self.f2t[0, facets]].flatten()
+        edge_candidates = np.unique(self.t2e[:, self.f2t[0, facets]])
         return edge_candidates[
             (self.edges.T[edge_candidates, None] == boundary_edges)
             .all(-1).any(-1)
