+++ skfem/version.py
@@ -1,1 +1,1 @@
-__version__ = "2.1.0"
+__version__ = "2.1.1.dev0"
