+++ README.md
@@ -182,6 +182,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Unreleased
 
+### [2.1.0] - 2020-09-30
+
 ### Fixed
 
 - `Mesh3D.boundary_edges` (and, consequently, `Basis.find_dofs`) was slow
