### 2020-10-01 d4aa46d
regenerate ex13.png on new structured mesh
