+++ docs/listofexamples.rst
@@ -161,7 +161,7 @@ y/x`, with :math:`u = 0` on :math:`y = 0`, :math:`u = 1` on :math:`x =
 0`, and :math:`\frac{\partial u}{\partial n} = 0` on the rest of the
 boundary.
 
-.. figure:: https://user-images.githubusercontent.com/973268/87775065-226f6d00-c82e-11ea-950c-fe9a10901133.png
+.. figure:: https://user-images.githubusercontent.com/1588947/94758580-d5d51000-03e0-11eb-8219-15cbba1d8c26.png
 
    The solution of Example 13.
 
