### 2020-10-01 a16124b
Remove ex13 from the list of exceptions
The dependency pygmsh was removed from ex13 in an earlier commit and, hence, we
relicense the file to BSD-3-Clause.
