### 2020-10-01 4f041e2
Optimize boundary_edges further and release 2.1.1
