+++ skfem/version.py
@@ -1,1 +1,1 @@
-__version__ = "2.1.1"
+__version__ = "2.1.2.dev0"
