+++ skfem/mesh/mesh_line.py
@@ -165,6 +165,17 @@ class MeshLine(Mesh):
         from skfem.mapping import MappingAffine
         return MappingAffine(self)
 
+    def element_finder(self, mapping=None):
+        ix = np.argsort(self.p)
+
+        def finder(x):
+            maxix = (x == np.max(self.p))
+            x[maxix] = x[maxix] - 1e-10  # special case in np.digitize
+            return np.argmax(np.digitize(x, self.p[0, ix[0]])[:, None]
+                             == self.t[0], axis=1)
+
+        return finder
+
     @staticmethod
     def strip_extra_coordinates(p: ndarray) -> ndarray:
         return p[:, :1]
