+++ skfem/version.py
@@ -1,1 +1,1 @@
-__version__ = "2.1.2.dev0"
+__version__ = "2.2.0.dev0"
