### 2020-10-06 4866e9f
Add action to run tests
