+++ .github/workflows/main.yml
@@ -1,0 +1,11 @@
+on: [push]
+
+jobs:
+  run_tests:
+    runs-on: ubuntu-latest
+    name: Run scikit-fem tests that use external packages
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+      - name: Hello world action step
+        uses: kinnala/scikit-fem-docker-action@v1
