+++ .github/workflows/main.yml
@@ -3,9 +3,9 @@ on: [push]
 jobs:
   run_tests:
     runs-on: ubuntu-latest
-    name: Run scikit-fem tests that use external packages
+    name: Run tests
     steps:
       - name: Checkout
         uses: actions/checkout@v2
-      - name: Hello world action step
-        uses: kinnala/scikit-fem-docker-action@v1
+      - name: Run tests inside a container
+        uses: kinnala/scikit-fem-docker-action@v2
