### 2020-10-06 c5dde4f
Support pygmsh 7.0.0 in examples
