+++ docs/examples/ex32.py
@@ -66,6 +66,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 """
 from typing import NamedTuple, Optional
+from packaging import version
 
 from skfem import *
 from skfem.io.meshio import from_meshio
