### 2020-10-06 d86c435
More work on pygmsh7 compatibility
