+++ docs/examples/ex32.py
@@ -67,6 +67,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 from typing import NamedTuple, Optional
 from packaging import version
+from contextlib import nullcontext
 
 from skfem import *
 from skfem.io.meshio import from_meshio
@@ -78,10 +79,13 @@ from scipy.sparse import bmat, spmatrix
 from scipy.sparse.linalg import LinearOperator, minres
 
 import pygmsh
+
+
 if version.parse(pygmsh.__version__) < version.parse('7.0.0'):
-    from pygmsh.opencascade import Geometry
+    geometrycontext = nullcontext()
 else:
-    from pygmsh.geo import Geometry
+    geometrycontext = pygmsh.occ.Geometry()
+
 
 try:
     try:
@@ -111,17 +115,19 @@ class Ellipsoid(NamedTuple):
     def semiaxes(self) -> np.ndarray:
         return np.array([self.a, self.b, self.c])
 
-    def geom(self, lcar: float) -> Geometry:
-        geom = Geometry()
-        geom.add_ellipsoid([0.]*3, self.semiaxes, lcar)
-        return geom
-
-    def mesh(self, geom: Optional[Geometry] = None, **kwargs) -> MeshTet:
-        if version.parse(pygmsh.__version__) < version.parse('7.0.0'):
-            return from_meshio(
-                pygmsh.generate_mesh(geom or self.geom(**kwargs)))
-        else:
-            return from_meshio((geom or self.geom(**kwargs)).generate_mesh())
+    def mesh(self, geom=None, lcar=.1) -> MeshTet:
+        with geometrycontext as g:
+            if version.parse(pygmsh.__version__) < version.parse('7.0.0'):
+                geom = pygmsh.opencascade.Geometry()
+            else:
+                geom = g
+            geom.add_ellipsoid([0.]*3, self.semiaxes, lcar)
+            if version.parse(pygmsh.__version__) < version.parse('7.0.0'):
+                return from_meshio(
+                    pygmsh.generate_mesh(geom))
+            else:
+                return from_meshio(
+                    geom.generate_mesh())
 
     def pressure(self, x, y, z) -> np.ndarray:
         """Exact pressure at zero Grashof number.
