### 2020-10-06 3eceb9b
Remove pygmsh from example 35
