+++ docs/examples/ex35.py
@@ -1,10 +1,6 @@
 """Computation of the characteristic impedance and velocity factor of RG316
 coaxial cable.
 
-.. note::
-   This example requires the external package
-   `pygmsh <https://pypi.org/project/pygmsh/>`_.
-
 This example solves the series inductance (per meter) and parallel capacitance
 (per meter)
 of RG316 coaxial cable. These values are then used to compute the
@@ -221,140 +217,21 @@ potential difference across the capacitor. Thus
 .. math::
    C = \\frac{2 E}{V^2}.
 
-License
--------
-
-Copyright 2018-2020 scikit-fem developers
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
 """
 from packaging import version
-from contextlib import nullcontext
+from pathlib import Path
 
 from skfem.mesh import MeshTri
 from skfem.assembly import InteriorBasis, FacetBasis
 from skfem.utils import solve, asm, condense, project
 from skfem.element import ElementTriP1
 from skfem.models.poisson import laplace, unit_load, mass
-from skfem.io import from_meshio
+from skfem.io.json import from_file
 
-import pygmsh
 import numpy as np
 
 
-if version.parse(pygmsh.__version__) < version.parse('7.0.0'):
-    geometrycontext = nullcontext()
-else:
-    geometrycontext = pygmsh.geo.Geometry()
-
-if version.parse(pygmsh.__version__) < version.parse('7.0.0'):
-
-    def make_mesh() -> MeshTri:
-        geom = pygmsh.built_in.Geometry()
-
-        # dimensions for RG316 coaxial cable
-        inner_conductor_diameter = 0.50e-3
-        inner_insulator_diameter = 1.52e-3
-        outer_conductor_diameter = 1.98e-3
-        outer_insulator_diameter = 2.48e-3
-
-        # characteristic length for mesh generation
-        lcar = 0.1e-3
-
-        inner_conductor = geom.add_circle(
-            (0, 0, 0), inner_conductor_diameter/2,
-            lcar)
-        geom.add_physical(
-            inner_conductor.plane_surface, label='inner_conductor')
-        geom.add_physical(
-            inner_conductor.line_loop.lines, label='inner_conductor_outer_surface')
-
-        inner_insulator = geom.add_circle(
-            (0, 0, 0), inner_insulator_diameter/2,
-            lcar, holes=[inner_conductor.line_loop])
-        geom.add_physical(
-            inner_insulator.plane_surface, label='inner_insulator')
-        geom.add_physical(
-            inner_insulator.line_loop.lines, label='outer_conductor_inner_surface')
-
-        outer_conductor = geom.add_circle(
-            (0, 0, 0), outer_conductor_diameter/2,
-            lcar, holes=[inner_insulator.line_loop])
-        geom.add_physical(
-            outer_conductor.plane_surface, label='outer_conductor')
-        geom.add_physical(
-            outer_conductor.line_loop.lines, label='outer_conductor_outer_surface')
-
-        outer_insulator = geom.add_circle(
-            (0, 0, 0), outer_insulator_diameter/2,
-            lcar, holes=[outer_conductor.line_loop])
-        geom.add_physical(
-            outer_insulator.plane_surface, label='outer_insulator')
[…]
