### 2020-10-06 ba53353
Load mesh from file as a temporary fix to ex28
