### 2020-10-09 acd449e
Small improvements to matplotlib plotting and bump version
