### 2020-10-10 5dc55d5
Modify wording in changelog
