+++ skfem/mesh/mesh2d/__init__.py
@@ -1,3 +1,4 @@
 from .mesh2d import Mesh2D  # noqa
 from .mesh_tri import MeshTri  # noqa
+from .mesh_tri_2 import MeshTri2  # noqa
 from .mesh_quad import MeshQuad  # noqa
