+++ skfem/mesh/__init__.py
@@ -35,7 +35,7 @@ Importing from external formats can be done with the constructor
 
 from .mesh import Mesh, MeshType
 from .mesh_line import MeshLine
-from .mesh2d import Mesh2D, MeshTri, MeshQuad
+from .mesh2d import Mesh2D, MeshTri, MeshQuad, MeshTri2
 from .mesh3d import Mesh3D, MeshTet, MeshHex
 
 
@@ -45,6 +45,7 @@ __all__ = [
     "MeshLine",
     "Mesh2D",
     "MeshTri",
+    "MeshTri2",
     "MeshQuad",
     "Mesh3D",
     "MeshTet",
