### 2020-10-11 ac95067
Simplify constructor
