+++ skfem/mesh/__init__.py
@@ -35,7 +35,8 @@ Importing from external formats can be done with the constructor
 
 from .mesh import Mesh, MeshType
 from .mesh_line import MeshLine
-from .mesh2d import Mesh2D, MeshTri, MeshQuad, MeshTri2
+from .mesh2d import (Mesh2D, MeshTri, MeshQuad,
+                     MeshTri2, MeshQuad2)
 from .mesh3d import Mesh3D, MeshTet, MeshHex
 
 
@@ -47,6 +48,7 @@ __all__ = [
     "MeshTri",
     "MeshTri2",
     "MeshQuad",
+    "MeshQuad2",
     "Mesh3D",
     "MeshTet",
     "MeshHex"]
