### 2020-10-11 630f45e
Move meshes to a separate directory
