+++ docs/examples/ex21.py
@@ -67,7 +67,7 @@ import numpy as np
 
 from pathlib import Path
 
-m = MeshTet.load(Path(__file__).with_name("beams.msh"))
+m = MeshTet.load(Path(__file__).parent / 'meshes' / 'beams.msh')
 e1 = ElementTetP2()
 e = ElementVectorH1(e1)
 
