### 2020-10-11 e922410
Add simple tests for quadratic meshes
