### 2020-10-11 0a03bb6
Support basic refine and simplify
