### 2020-10-12 9ea44c0
Start adding mypy support
