+++ skfem/assembly/form/linear_form.py
@@ -16,30 +16,30 @@ class LinearForm(Form):
     """
 
     def assemble(self,
-                 u: Basis,
-                 v: Optional[Basis] = None,
+                 u_basis: Basis,
+                 v_basis: Optional[Basis] = None,
                  **kwargs) -> ndarray:
 
-        assert v is None
-        v = u
+        assert v_basis is None
+        v_basis = u_basis
 
-        nt = v.nelems
-        dx = v.dx
-        w = FormDict({**v.default_parameters(), **self.dictify(kwargs)})
+        nt = v_basis.nelems
+        dx = v_basis.dx
+        w = FormDict({**v_basis.default_parameters(), **self.dictify(kwargs)})
 
         # initialize COO data structures
-        sz = v.Nbfun * nt
+        sz = v_basis.Nbfun * nt
         data = np.zeros(sz, dtype=self.dtype)
         rows = np.zeros(sz)
         cols = np.zeros(sz)
 
-        for i in range(v.Nbfun):
+        for i in range(v_basis.Nbfun):
             ixs = slice(nt * i, nt * (i + 1))
-            rows[ixs] = v.element_dofs[i]
+            rows[ixs] = v_basis.element_dofs[i]
             cols[ixs] = np.zeros(nt)
-            data[ixs] = self._kernel(v.basis[i], w, dx)
+            data[ixs] = self._kernel(v_basis.basis[i], w, dx)
 
-        return self._assemble_numpy_vector(data, rows, cols, (v.N, 1))
+        return self._assemble_numpy_vector(data, rows, cols, (v_basis.N, 1))
 
     def _kernel(self, v, w, dx):
         return np.sum(self.form(*v, w) * dx, axis=1)
