+++ skfem/assembly/form/linear_form.py
@@ -16,30 +16,30 @@ class LinearForm(Form):
     """
 
     def assemble(self,
-                 u_basis: Basis,
-                 v_basis: Optional[Basis] = None,
+                 ubasis: Basis,
+                 vbasis: Optional[Basis] = None,
                  **kwargs) -> ndarray:
 
-        assert v_basis is None
-        v_basis = u_basis
+        assert vbasis is None
+        vbasis = ubasis
 
-        nt = v_basis.nelems
-        dx = v_basis.dx
-        w = FormDict({**v_basis.default_parameters(), **self.dictify(kwargs)})
+        nt = vbasis.nelems
+        dx = vbasis.dx
+        w = FormDict({**vbasis.default_parameters(), **self.dictify(kwargs)})
 
         # initialize COO data structures
-        sz = v_basis.Nbfun * nt
+        sz = vbasis.Nbfun * nt
         data = np.zeros(sz, dtype=self.dtype)
         rows = np.zeros(sz)
         cols = np.zeros(sz)
 
-        for i in range(v_basis.Nbfun):
+        for i in range(vbasis.Nbfun):
             ixs = slice(nt * i, nt * (i + 1))
-            rows[ixs] = v_basis.element_dofs[i]
+            rows[ixs] = vbasis.element_dofs[i]
             cols[ixs] = np.zeros(nt)
-            data[ixs] = self._kernel(v_basis.basis[i], w, dx)
+            data[ixs] = self._kernel(vbasis.basis[i], w, dx)
 
-        return self._assemble_numpy_vector(data, rows, cols, (v_basis.N, 1))
+        return self._assemble_numpy_vector(data, rows, cols, (vbasis.N, 1))
 
     def _kernel(self, v, w, dx):
         return np.sum(self.form(*v, w) * dx, axis=1)
