### 2020-10-12 a7f76d7
Rename functions
