### 2020-10-12 864efd9
Make type stable
