+++ skfem/io/json.py
@@ -6,13 +6,15 @@ from skfem.mesh import MeshLine, MeshTri, MeshQuad,\
     MeshTet, MeshHex, Mesh
 
 
-def from_file(filename: str):
+def from_file(filename: str) -> Mesh:
     with open(filename, 'r') as handle:
         d = json.load(handle)
 
     # detect dimension and number of vertices
     dim = len(d['p'][0])
     nverts = len(d['t'][0])
+
+    mesh_type = Mesh
 
     if dim == 1:
         mesh_type = MeshLine
@@ -32,4 +34,4 @@ def from_file(filename: str):
 
 def to_file(mesh: Mesh, filename: str):
     with open(filename, 'w') as handle:
-        _ = json.dump(mesh.to_dict(), handle)
+        json.dump(mesh.to_dict(), handle)
