### 2020-10-12 c97027a
Remove obsolete properties from DiscreteField
