+++ skfem/element/discrete_field.py
@@ -14,39 +14,18 @@ class DiscreteField(NamedTuple):
     hess: Optional[ndarray] = None
     hod: Optional[ndarray] = None
 
-    @property
-    def f(self):
-        """For backwards compatibility; used by old style form decorators."""
-        return self.value
-
-    @property
-    def df(self):
-        """For backwards compatibility; used by old style form decorators."""
-        if self.grad is not None:
-            return self.grad
-        elif self.div is not None:
-            return self.div
-        elif self.curl is not None:
-            return self.curl
-        return None
-
-    @property
-    def ddf(self):
-        """For backwards compatibility; used by old style form decorators."""
-        return self.hess
-
     def __array__(self):
-        return self.f
+        return self.value
 
     def __mul__(self, other):
         if isinstance(other, DiscreteField):
-            return self.f * other.f
-        return self.f * other
+            return self.value * other.value
+        return self.value * other
 
     def _split(self):
         """Split all components based on their first dimension."""
         return [DiscreteField(*[f[i] for f in self if f is not None])
-                for i in range(self.f.shape[0])]
+                for i in range(self.value.shape[0])]
 
     def zeros_like(self):
         """Return zero :class:`~skfem.element.DiscreteField` with same size."""
@@ -58,4 +37,5 @@ class DiscreteField(NamedTuple):
 
         return DiscreteField(*[zero_or_none(field) for field in self])
 
-    __rmul__ = __mul__
+
+DiscreteField.__rmul__ = DiscreteField.__mul__
