+++ tests/test_convergence_hdiv.py
@@ -24,7 +24,7 @@ class ConvergenceRaviartThomas(unittest.TestCase):
 
         @BilinearForm
         def bilinf_B(sigma, v, w):
-            return sigma.df * v
+            return sigma.div * v
 
         @LinearForm
         def load(v, w):
@@ -98,7 +98,7 @@ class ConvergenceRaviartThomas(unittest.TestCase):
                     np.sin(np.pi * y[1]) *
                     np.sin(np.pi * y[2]))
 
-        return np.sqrt(np.sum(np.sum((uh - u(x.f)) ** 2 * dx, axis=1)))
+        return np.sqrt(np.sum(np.sum((uh - u(x.value)) ** 2 * dx, axis=1)))
 
     def compute_Hdiv(self, m, basis, U):
         uh, duh, *_ = basis.interpolate(U)
@@ -123,7 +123,7 @@ class ConvergenceRaviartThomas(unittest.TestCase):
 
         divuh = sum(uh)
 
-        return np.sqrt(np.sum(np.sum(((divuh - divu(x.f)) ** 2) * dx, axis=1)))
+        return np.sqrt(np.sum(np.sum(((divuh - divu(x.value)) ** 2) * dx, axis=1)))
 
     def create_basis(self, m):
         e = ElementTriRT0()
