+++ skfem/element/element_line/element_line_pp.py
@@ -1,4 +1,5 @@
 import warnings
+from typing import Type
 
 import numpy as np
 from numpy.polynomial.legendre import Legendre
@@ -11,7 +12,7 @@ class ElementLinePp(ElementH1):
 
     nodal_dofs = 1
     dim = 1
-    mesh_type = MeshLine
+    mesh_type: Type = MeshLine
 
     def __init__(self, p):
         if p < 1:
