+++ skfem/element/element_quad/element_quadp.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from ...mesh.mesh2d import MeshQuad
+from ...mesh import MeshQuad
 from ..element_line import ElementLinePp
 
 
