### 2020-10-12 2b8c3b5
Fix typing in skfem.mesh
