+++ skfem/mesh/mesh2d/mesh2d.py
@@ -1,7 +1,7 @@
 import numpy as np
 from numpy import ndarray
 
-from skfem.mesh import Mesh, MeshType
+from skfem.mesh import Mesh
 
 
 class Mesh2D(Mesh):
@@ -18,7 +18,10 @@ class Mesh2D(Mesh):
     f2t = np.zeros((2, 0), dtype=np.int64)
     t2f = np.array([], dtype=np.int64)
 
-    def mirror(self, a: float, b: float, c: float) -> MeshType:
+    def __init__(self, *args, **kwargs):
+        super(Mesh2D, self).__init__()
+
+    def mirror(self, a: float, b: float, c: float) -> Mesh:
         """Mirror a mesh by the line :math:`ax + by + c = 0`.  Returns a new
         :class:`~skfem.mesh.Mesh` object."""
         tmp = -2. * (a * self.p[0] + b * self.p[1] + c) / (a ** 2 + b ** 2)
