### 2020-10-12 4c09736
Fix rest of typing issues
