+++ skfem/element/element_composite.py
@@ -1,3 +1,5 @@
+from typing import List, Any
+
 import numpy as np
 from numpy import ndarray
 
@@ -54,8 +56,9 @@ class ElementComposite(Element):
 
     def _deduce_bfun(self, i: int):
         """Deduce component and basis function for i'th index."""
-        counts = sum([e._bfun_counts() for e in self.elems])
-        ns = []
+        counts: ndarray = sum([e._bfun_counts() for e in self.elems])
+        tmp: List[Any] = []
+        ns: List[Any] = []
         if counts[0] > 0:
             tmp = sum([[j] * self.elems[j].nodal_dofs
                        for j in range(len(self.elems))], [])
