### 2020-10-12 48873a4
Readd DiscreteField.f/df/ddf
