### 2020-10-13 a591c3a
Add information about testing
