### 2020-10-13 e096118
Add few warnings about deprecations
