### 2020-10-13 fd4ee04
Do not expand facet DOFs unless it's required
