### 2020-10-13 f431442
Remove Travis badge
