+++ skfem/element/element_tet/element_tet_cr.py
@@ -1,0 +1,39 @@
+import numpy as np
+
+from ..element_h1 import ElementH1
+from ...mesh.mesh3d import MeshTet
+
+
+class ElementTetCR(ElementH1):
+
+    facet_dofs = 1
+    dim = 3
+    maxdeg = 1
+    dofnames = ['u']
+    doflocs = np.array([[1 / 3, 1 / 3, 0.],
+                        [1 / 3, 0., 1 / 3],
+                        [0., 1 / 3, 1 / 3],
+                        [1 / 3, 1 / 3, 1 / 3]])
+    mesh_type = MeshTet
+
+    def lbasis(self, X, i):
+        x, y, z = X
+
+        if i == 0:
+            phi = 1. - 3. * z
+            dphi = np.array([0. * x, 0. * y, -3. + 0. * z])
+        elif i == 1:
+            phi = 1. - 3. * y
+            dphi = np.array([0. * x, -3. + 0. * y, 0. * z])
+        elif i == 2:
+            phi = 1. - 3. * x
+            dphi = np.array([-3. + 0. * x, 0. * y, 0. * z])
+        elif i == 3:
+            phi = 3. * x + 3. * y + 3. * z - 2.
+            dphi = np.array([3. + 0. * x,
+                             3. + 0. * y,
+                             3. + 0. * z])
+        else:
+            self._index_error()
+
+        return phi, dphi
