+++ skfem/element/element_tet/__init__.py
@@ -4,3 +4,4 @@ from .element_tet_p2 import ElementTetP2  # noqa
 from .element_tet_rt0 import ElementTetRT0  # noqa
 from .element_tet_n0 import ElementTetN0  # noqa
 from .element_tet_mini import ElementTetMini  # noqa
+from .element_tet_cr import ElementTetCR  # noqa
