### 2020-10-14 f91ee90
Run tests also in Python 3.8
