+++ .github/workflows/main.yml
@@ -3,7 +3,7 @@ on: [push, pull_request]
 jobs:
   run_tests_36:
     runs-on: ubuntu-latest
-    name: Run tests on Python 3.6
+    name: Python 3.6
     steps:
       - name: Checkout
         uses: actions/checkout@v2
@@ -11,7 +11,7 @@ jobs:
         uses: kinnala/scikit-fem-docker-action@py36
   run_tests_37:
     runs-on: ubuntu-latest
-    name: Run tests on Python 3.7
+    name: Python 3.7
     steps:
       - name: Checkout
         uses: actions/checkout@v2
@@ -19,17 +19,9 @@ jobs:
         uses: kinnala/scikit-fem-docker-action@py37
   run_tests_38:
     runs-on: ubuntu-latest
-    name: Run tests on Python 3.8
+    name: Python 3.8
     steps:
       - name: Checkout
         uses: actions/checkout@v2
       - name: Run tests
         uses: kinnala/scikit-fem-docker-action@py38
-  run_tests_39:
-    runs-on: ubuntu-latest
-    name: Run tests on Python 3.9
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Run tests
-        uses: kinnala/scikit-fem-docker-action@py39
