### 2020-10-14 dbfe7a8
Add options for running the different versions in Makefile
