+++ skfem/__init__.py
@@ -5,7 +5,6 @@ from skfem.assembly import *  # noqa
 from skfem.mapping import *  # noqa
 from skfem.element import *  # noqa
 from skfem.utils import *  # noqa
-from skfem.version import __version__  # noqa
 
 from skfem.assembly import __all__ as all_assembly
 from skfem.mesh import __all__ as all_mesh
