+++ setup.cfg
@@ -1,0 +1,36 @@
+[metadata]
+name = scikit-fem
+version = 2.2.0.dev0
+author = Tom Gustafsson
+description = Simple finite element assemblers
+url = https://github.com/kinnala/scikit-fem
+project_urls =
+    Code=https://github.com/kinnala/scikit-fem
+    Issues=https://github.com/kinnala/scikit-fem/issues
+long_description = file: README.md
+long_description_content_type = text/markdown
+license = BSD-3-Clause
+license_file = LICENSE.md
+classifiers =
+    Intended Audience :: Science/Research
+    Operating System :: OS Independent
+    License :: OSI Approved :: BSD License
+    Programming Language :: Python
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Topic :: Scientific/Engineering
+    Topic :: Utilities
+
+[options]
+packages = find:
+install_requires =
+    importlib_metadata;python_version<"3.8"
+    numpy
+    scipy
+    meshio>=4.0.4
+python_requires = >=3.6
+setup_requires =
+    setuptools>=42
+    wheel
