### 2020-10-14 6fb1002
Support editable mode via 'python _setup.py develop'
