### 2020-10-14 685c08b
Update changelog and fix Makefile
