+++ README.md
@@ -192,6 +192,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Unreleased
 
+### [2.2.0] - 2020-10-14
+
 ### Fixed
 
 - Fix `Mesh.validate` for unsigned `Mesh.t`.
