### 2020-10-14 f1c404a
Do not include tests in PyPI, bump up version number
