+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = scikit-fem
-version = 2.2.0
+version = 2.2.1
 author = Tom Gustafsson
 description = Simple finite element assemblers
 url = https://github.com/kinnala/scikit-fem
@@ -34,3 +34,7 @@ python_requires = >=3.6
 setup_requires =
     setuptools>=42
     wheel
+
+[options.packages.find]
+exclude =
+    tests
