+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = scikit-fem
-version = 2.2.1
+version = 2.2.2
 author = Tom Gustafsson
 description = Simple finite element assemblers
 url = https://github.com/kinnala/scikit-fem
