### 2020-10-15 541d99e
Add entires and make changelog more compact
