+++ README.md
@@ -192,119 +192,93 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Unreleased
 
+### [2.2.2] - 2020-10-15
+
+- Fixed: Make the preconditioner in `TestEx32` more robust.
+
+### [2.2.1] - 2020-10-15
+
+- Fixed: Remove `tests` from the PyPI distribution.
+
 ### [2.2.0] - 2020-10-14
 
-### Fixed
-
-- Fix `Mesh.validate` for unsigned `Mesh.t`.
-
-### Added
-
-- `MeshTet.element_finder` and `MeshLine.element_finder` for using
-  `InteriorBasis.interpolator`.
-- `ElementTriCR`, the nonconforming Crouzeix-Raviart element for Stokes flow.
-- `ElementTetCR`, tetrahedral nonconforming Crouzeix-Raviart element.
-- `ElementTriHermite`, an extension of `ElementLineHermite` to triangular
+- Fixed: Fix `Mesh.validate` for unsigned `Mesh.t`.
+- Added: `MeshTet.element_finder` and `MeshLine.element_finder` for using
+  Added: `InteriorBasis.interpolator`.
+- Added: `ElementTriCR`, the nonconforming Crouzeix-Raviart element for Stokes flow.
+- Added: `ElementTetCR`, tetrahedral nonconforming Crouzeix-Raviart element.
+- Added: `ElementTriHermite`, an extension of `ElementLineHermite` to triangular
   meshes.
-
-### Deprecated
-
-- `L2_projection` will be replaced by `project`.
-- `derivative` will be replaced by `project`.
+- Deprecated: `L2_projection` will be replaced by `project`.
+- Deprecated: `derivative` will be replaced by `project`.
 
 ### [2.1.1] - 2020-10-01
 
-### Fixed
-
-- Further optimizations to `Mesh3D.boundary_edges`: tested to run on a laptop
+- Fixed: Further optimizations to `Mesh3D.boundary_edges`: tested to run on a laptop
   with over 10 million elements.
 
 ### [2.1.0] - 2020-09-30
 
-### Fixed
-
-- `Mesh3D.boundary_edges` (and, consequently, `Basis.find_dofs`) was slow
+- Fixed: `Mesh3D.boundary_edges` (and, consequently, `Basis.find_dofs`) was slow
   and used lots of memory due to an exhaustive search of all edges
-
-#### Added
-- `ElementHex2`, a triquadratic hexahedral element
-- `MeshTri.init_circle`, constructor for a circle mesh
+- Added: `ElementHex2`, a triquadratic hexahedral element
+- Added: `MeshTri.init_circle`, constructor for a circle mesh
 
 ### [2.0.0] - 2020-08-21
 
-#### Added
-- Support for complex-valued forms: `BilinearForm` and `LinearForm` now take
+- Added: Support for complex-valued forms: `BilinearForm` and `LinearForm` now take
   an optional argument `dtype` which defaults to `np.float64`
   but can be also `np.complex64`
-- `Dofs.__or__` and `Dofs.__add__`, for merging degree-of-freedom sets
+- Added: `Dofs.__or__` and `Dofs.__add__`, for merging degree-of-freedom sets
   (i.e. `Dofs` objects) using `|` and `+` operators
-- `Dofs.drop` and `Dofs.keep`, for further filtering the degree-of-freedom sets
-
-### Removed
-- Support for old-style decorators `bilinear_form`, `linear_form`, and
+- Added: `Dofs.drop` and `Dofs.keep`, for further filtering the degree-of-freedom sets
[…]
