### 2020-10-15 15546b2
Fix typo and add information about release testing
