+++ README.md
@@ -192,6 +192,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Unreleased
 
+### [2.2.3] - 2020-10-16
+
+- Fixed: Remove an unnecessary dependency.
+
 ### [2.2.2] - 2020-10-15
 
 - Fixed: Make the preconditioner in `TestEx32` more robust.
