+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = scikit-fem
-version = 2.2.2
+version = 2.2.3
 author = Tom Gustafsson
 description = Simple finite element assemblers
 url = https://github.com/kinnala/scikit-fem
@@ -26,7 +26,6 @@ classifiers =
 [options]
 packages = find:
 install_requires =
-    importlib_metadata;python_version<"3.8"
     numpy
     scipy
     meshio>=4.0.4
