### 2020-10-16 c7a4fdc
Add MeshTet.init_sphere and remove pygmsh dependency from ex32
