### 2020-10-16 c3ef537
Rename init_sphere to init_ball
