### 2020-10-19 a63f3f9
don't stretch ball into ellipsoid in ex32 #498
