+++ docs/examples/ex32.py
@@ -92,7 +92,7 @@ class Sphere(NamedTuple):
 
         """
 
-        a, b, c = np.ones(3)
+        a, b, _ = np.ones(3)  # semiaxes of ellipsoid
         return (a**2 * (3 * a**2 + b**2) * x * y
                 / (3 * a**4 + 2 * a**2 * b**2 + 3 * b**4))
 
