### 2020-10-19 b781fe3
new figure for ex32 in a sphere
