### 2020-10-20 e3beb90
DOC: update meshing of ex13, forgotten in #486
