+++ docs/examples/ex13.py
@@ -2,16 +2,14 @@ r"""Laplace with mixed boundary conditions.
 
 This example is another extension of `ex01.py`, still solving the Laplace
 equation but now with mixed boundary conditions, two parts isopotential (charged
-and earthed) and the rest insulated. The isopotential parts are tagged during
-the construction of the geometry in `pygmsh
-<https://pypi.org/project/pygmsh/>`_.
+and earthed) and the rest insulated.
 
 The example is :math:`\Delta u = 0` in
 :math:`\Omega=\{(x,y):1<x^2+y^2<4,~0<\theta<\pi/2\}`, where :math:`\tan \theta =
 y/x`, with :math:`u = 0` on :math:`y = 0` and :math:`u = 1` on :math:`x =
-0`. Although these boundaries would be simple enough to identify using the
-coordinates and :meth:`skfem.assembly.Basis.find_dofs`, the present
-technique generalizes to more complicated shapes.
+0`.  The mesh is first constructed as a rectangle in the :math:`r\theta`-plane,
+where the isopotential parts are conveniently tagged using `skfem.Mesh.define_boundary`;
+then the mesh is mapped to the :math:`xy`-plane.
 
 The exact solution is :math:`u = 2 \theta / \pi`. The field strength is :math:`|\nabla u|^2 = 4 / \pi^2 (x^2 + y^2)`
 so the conductance (for unit potential difference and conductivity) is
