### 2020-10-20 4b9bbd0
DOC: ex35 no longer uses pygmsh #489
