### 2020-10-20 0f66d04
DOC: ex04 no longer uses pygmsh #493
