### 2020-10-27 aeb61ff
Give a name to the action and update README
