+++ README.md
@@ -46,6 +46,12 @@ of `scipy.sparse.linalg.spsolve`.*
 | 512000 | 8.82814 | nan |
 | 1030301 | 18.25461 | nan |
 
+## Installation
+
+[![PyPI](https://img.shields.io/pypi/v/scikit-fem)](https://pypi.org/project/scikit-fem/)
+[![Conda](https://img.shields.io/conda/vn/conda-forge/scikit-fem)](https://anaconda.org/conda-forge/scikit-fem)
+
+The most recent release can be installed simply by `pip install scikit-fem`.
 
 ## Examples
 
@@ -88,24 +94,22 @@ More examples can be found in the [gallery](https://scikit-fem.readthedocs.io/en
 
 ## Documentation
 
-The project is documented using Sphinx.  A recent version of the documentation
-can be found from [Read the Docs](https://scikit-fem.readthedocs.io/en/latest/).
+[![Documentation](https://readthedocs.org/projects/pip/badge/?version=stable)](https://scikit-fem.readthedocs.io/)
+
+The project is documented using Sphinx under `/docs`.  Built version of the
+documentation can be found from [Read the
+Docs](https://scikit-fem.readthedocs.io/en/latest/).
 
 ## Getting help
 
 If you encounter an issue and cannot find help from the documentation,
-you can use the Github issue tracker to ask questions.
+you can use the Github issue tracker to [ask questions using the question label](https://github.com/kinnala/scikit-fem/issues?q=label%3Aquestion).
 Try to provide a snippet of code which fails
 and include also the version of the library you are
 using.  The version can be found as follows:
 ```
 python -c "import pkg_resources; print(pkg_resources.get_distribution('scikit-fem').version)"
 ```
-
-## Installation
-
-The most recent release can be installed simply by `pip install scikit-fem`.
-For more cutting edge features, you can clone this repository.
 
 ## Dependencies
 
@@ -114,20 +118,24 @@ The minimal dependencies for installing `scikit-fem` are
 [meshio](https://github.com/nschloe/meshio).  In addition, many
 [examples](https://scikit-fem.readthedocs.io/en/latest/listofexamples.html) use
 [matplotlib](https://matplotlib.org/) for visualization.  Some examples
-demonstrate the use of other external packages; see our [CI job
-definition](https://github.com/kinnala/scikit-fem/blob/master/.github/workflows/main.yml) for a
-full list of test dependencies.
+demonstrate the use of other external packages; see `requirements.txt` for a
+list of test dependencies.
 
 ## Testing
 
+[![Tests](https://github.com/kinnala/scikit-fem/workflows/tests/badge.svg)](https://github.com/kinnala/scikit-fem/actions)
+[![Release tests](https://github.com/kinnala/scikit-fem-release-tests/workflows/release%20tests/badge.svg)](https://github.com/kinnala/scikit-fem-release-tests/actions)
+
 The tests are run by Github Actions.  The `Makefile` in the repository root has
-instructions for running the testing container locally using `docker`.  For
-example, use `make test_py38` to run the tests using `py38` branch from
+targets for running the testing container locally using `docker`.  For example,
+`make test_py38` runs the tests using `py38` branch from
 [kinnala/scikit-fem-docker-action](https://github.com/kinnala/scikit-fem-docker-action).
 The releases are tested in
 [kinnala/scikit-fem-release-tests](https://github.com/kinnala/scikit-fem-release-tests).
 
 ## Licensing
+
+[![License](https://img.shields.io/badge/license-BSD%203--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause)
 
[…]
