+++ .github/workflows/main.yml
@@ -1,3 +1,4 @@
+name: tests
 on: [push, pull_request]
 
 jobs:
