### 2020-10-27 d8efb84
Add funding body and Zenodo badge
