### 2020-10-27 3d5b71c
Some changes to badges, e.g., add versions
