+++ README.md
@@ -3,15 +3,15 @@
 </p>
 
 <p align="center">
-[![PyPI](https://img.shields.io/pypi/v/scikit-fem)](https://pypi.org/project/scikit-fem/)
-[![PyPI - Python Version](https://img.shields.io/pypi/pyversions/scikit-fem)](https://pypi.org/project/scikit-fem/)
-[![Conda](https://img.shields.io/conda/vn/conda-forge/scikit-fem)](https://anaconda.org/conda-forge/scikit-fem)
-[![Documentation](https://readthedocs.org/projects/pip/badge/?version=stable)](https://scikit-fem.readthedocs.io/)
-[![License](https://img.shields.io/badge/license-BSD%203--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause)
-[![Tests](https://github.com/kinnala/scikit-fem/workflows/tests/badge.svg)](https://github.com/kinnala/scikit-fem/actions)
-[![Release tests](https://github.com/kinnala/scikit-fem-release-tests/workflows/release%20tests/badge.svg)](https://github.com/kinnala/scikit-fem-release-tests/actions)
-[![status](https://joss.theoj.org/papers/4120aba1525403e6d0972f4270d7b61e/status.svg)](https://joss.theoj.org/papers/4120aba1525403e6d0972f4270d7b61e)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1420510.svg)](https://doi.org/10.5281/zenodo.1420510)
+<a href="https://pypi.org/project/scikit-fem/" alt="PyPI"><img src="https://img.shields.io/pypi/v/scikit-fem" /></a>
+<a href="https://pypi.org/project/scikit-fem/" alt="PyPI - Python Version"><img src="https://img.shields.io/pypi/pyversions/scikit-fem" /></a>
+<a href="https://anaconda.org/conda-forge/scikit-fem" alt="Conda"><img src="https://img.shields.io/conda/vn/conda-forge/scikit-fem" /></a>
+<a href="https://scikit-fem.readthedocs.io/" alt="Documentation"><img src="https://readthedocs.org/projects/pip/badge/?version=stable" /></a>
+<a href="https://opensource.org/licenses/BSD-3-Clause" alt="License"><img src="https://img.shields.io/badge/license-BSD%203--Clause-blue.svg" /></a>
+<a href="https://github.com/kinnala/scikit-fem/actions" alt="Tests"><img src="https://github.com/kinnala/scikit-fem/workflows/tests/badge.svg" /></a>
+<a href="https://github.com/kinnala/scikit-fem-release-tests/actions" alt="Release tests"><img src="https://github.com/kinnala/scikit-fem-release-tests/workflows/release%20tests/badge.svg" /></a>
+<a href="https://joss.theoj.org/papers/4120aba1525403e6d0972f4270d7b61e" alt="status"><img src="https://joss.theoj.org/papers/4120aba1525403e6d0972f4270d7b61e/status.svg" /></a>
+<a href="https://doi.org/10.5281/zenodo.1420510" alt="DOI"><img src="https://zenodo.org/badge/DOI/10.5281/zenodo.1420510.svg" /></a>
 </p>
 
 
