### 2020-11-05 9b6dc60
vrot: stream-function -> velocity
