+++ docs/examples/ex20.py
@@ -52,30 +52,29 @@ stokes = asm(biharmonic, ib)
 rotf = asm(unit_load, ib)
 
 psi = solve(*condense(stokes, rotf, D=ib.find_dofs()))
-psi0, = ib.interpolator(psi)(np.zeros((2, 1)))
+(psi0,) = ib.interpolator(psi)(np.zeros((2, 1)))
 
 if __name__ == "__main__":
     from os.path import splitext
     from sys import argv
+    from skfem.models.general import vrot
     from skfem.visuals.matplotlib import draw
     from matplotlib.tri import Triangulation
 
-    print('psi0 = {} (cf. exact = 1/64 = {})'.format(psi0, 1/64))
+    print("psi0 = {} (cf. exact = 1/64 = {})".format(psi0, 1 / 64))
 
     M, Psi = ib.refinterp(psi, 3)
 
     ax = draw(mesh)
     ax.tricontour(Triangulation(*M.p, M.t.T), Psi)
     name = splitext(argv[0])[0]
-    ax.get_figure().savefig(f'{name}_stream-lines.png')
+    ax.get_figure().savefig(f"{name}_stream-lines.png")
 
-    refbasis = InteriorBasis(M, ElementTriP1())
-    velocity = np.vstack([derivative(Psi, refbasis, refbasis, 1),
-                          -derivative(Psi, refbasis, refbasis, 0)])
+    velocity = asm(
+        vrot,
+        InteriorBasis(mesh, ElementVectorH1(ElementTriP1()), quadrature=ib.quadrature),
+        w=ib.interpolate(psi),
+    )
     ax = draw(mesh)
-    sparsity_factor = 2**3      # subsample the arrows
-    vector_factor = 2**3        # lengthen the arrows
-    x = M.p[:, ::sparsity_factor]
-    u = vector_factor * velocity[:, ::sparsity_factor]
-    ax.quiver(*x, *u, x[0])
-    ax.get_figure().savefig(f'{name}_velocity-vectors.png')
+    ax.quiver(*mesh.p, *velocity, mesh.p[0])
+    ax.get_figure().savefig(f"{name}_velocity-vectors.png")
