### 2020-11-06 fe14240
use new implementations of two-dimensional curl from #507
