+++ skfem/models/general.py
@@ -1,9 +1,7 @@
 """Bilinear and linear forms too general to put into a specific model."""
 
 from skfem.assembly import BilinearForm, LinearForm
-from .helpers import div, grad
-
-import numpy as np
+from .helpers import dot, div, curl
 
 
 @BilinearForm
@@ -13,11 +11,9 @@ def divergence(u, v, w):
 
 @LinearForm
 def rot(v, w):
-    return np.einsum('i...,ij...,j...',
-                     w.w, np.array([[0, 1], [-1, 0]]), grad(v))
+    return dot(curl(v), w.w)
 
 
 @LinearForm
 def vrot(v, w):
-    return np.einsum('i...,ij...,j...',
-                     v, np.array([[0, 1], [-1, 0]]), grad(w['w']))
+    return dot(v, curl(w.w))
