### 2020-11-06 e4c0a4f
Add changelog entry and change badge order
