+++ tests/test_elements.py
@@ -3,15 +3,15 @@ from unittest import TestCase, main
 import numpy as np
 from numpy.testing import assert_allclose, assert_array_equal
 
-from skfem.element import (ElementHex1, ElementHexS2, ElementLineP1, 
-                           ElementLineP2, ElementLinePp, ElementLineMini,
-                           ElementQuad0, ElementQuad1, ElementQuad2,
-                           ElementQuadP, ElementQuadS2, ElementTetMini, 
-                           ElementTetP0, ElementTetP1, ElementTetP2,
-                           ElementTriMini, ElementTriP0, ElementTriP1,
-                           ElementTriP2, ElementTriRT0, ElementVectorH1,
-                           ElementHex2, ElementQuadBFS, ElementTriCR,
-                           ElementTetCR, ElementTriHermite)
+from skfem.element import (ElementHex1, ElementHexS2, ElementLineP0,
+                           ElementLineP1, ElementLineP2, ElementLinePp,
+                           ElementLineMini, ElementQuad0, ElementQuad1,
+                           ElementQuad2, ElementQuadP, ElementQuadS2,
+                           ElementTetMini, ElementTetP0, ElementTetP1,
+                           ElementTetP2, ElementTriMini, ElementTriP0,
+                           ElementTriP1, ElementTriP2, ElementTriRT0,
+                           ElementVectorH1, ElementHex2, ElementQuadBFS,
+                           ElementTriCR, ElementTetCR, ElementTriHermite)
 from skfem.mesh import MeshHex, MeshLine, MeshQuad, MeshTet, MeshTri
 from skfem.assembly.basis import InteriorBasis
 from skfem.mapping import MappingAffine
@@ -21,6 +21,7 @@ class TestNodality(TestCase):
     """Test for Element.doflocs."""
 
     elems = [
+        ElementLineP0(),
         ElementLineP1(),
         ElementLineP2(),
         ElementLinePp(1),
