+++ skfem/element/__init__.py
@@ -45,7 +45,7 @@ from .element_tet import (ElementTetP0, ElementTetP1, ElementTetP2,
                           ElementTetRT0, ElementTetN0, ElementTetMini,
                           ElementTetCR)
 from .element_hex import ElementHex1, ElementHex2, ElementHexS2  # noqa
-from .element_line import (ElementLineP1, ElementLineP2,
+from .element_line import (ElementLineP0, ElementLineP1, ElementLineP2,
                            ElementLinePp, ElementLineHermite,
                            ElementLineMini)  # noqa
 from .element_composite import ElementComposite  # noqa
