+++ skfem/element/element_line/__init__.py
@@ -1,3 +1,4 @@
+from .element_line_p0 import ElementLineP0  # noqa
 from .element_line_p1 import ElementLineP1  # noqa
 from .element_line_p2 import ElementLineP2  # noqa
 from .element_line_pp import ElementLinePp  # noqa
