### 2020-11-09 25a636a
export ElementLineP0 from skfem.element
