### 2020-11-09 b8d96fb
check i==0 for P0 elements
