+++ skfem/element/element_quad/element_quad0.py
@@ -14,4 +14,9 @@ class ElementQuad0(ElementH1):
     mesh_type = MeshQuad
 
     def lbasis(self, X, i):
-        return np.ones(X.shape[1]), np.zeros_like(X)
+        if i == 0:
+            return np.ones(X.shape[1]), np.zeros_like(X)
+        else:
+            self._index_error()
+
+                
