+++ skfem/element/element_tet/element_tet_p0.py
@@ -13,4 +13,7 @@ class ElementTetP0(ElementH1):
     mesh_type = MeshTet
 
     def lbasis(self, X, i):
-        return 1. + 0. * X[0], 0. * X
+        if i == 0:
+            return 1. + 0. * X[0], 0. * X
+        else:
+            self._index_error()
