+++ skfem/element/element_quad/element_quad0.py
@@ -18,5 +18,3 @@ class ElementQuad0(ElementH1):
             return np.ones(X.shape[1]), np.zeros_like(X)
         else:
             self._index_error()
-
-                
