+++ skfem/element/element_quad/element_quad0.py
@@ -15,6 +15,6 @@ class ElementQuad0(ElementH1):
 
     def lbasis(self, X, i):
         if i == 0:
-            return np.ones(X.shape[1]), np.zeros_like(X)
+            return np.ones(X.shape[1:]), np.zeros_like(X)
         else:
             self._index_error()
