### 2020-11-10 63ffd0e
Fix dofloc to be in the middle of the elemetn
