### 2020-11-10 cadd7e9
Make running doctest possible
