+++ skfem/element/element_hex/__init__.py
@@ -1,3 +1,4 @@
+"""Yadda"""
 from .element_hex1 import ElementHex1  # noqa
 from .element_hex_s2 import ElementHexS2  # noqa
 from .element_hex2 import ElementHex2  # noqa
