+++ docs/elements.rst
@@ -1,0 +1,7 @@
+=================
+ Finite elements
+=================
+
+.. automodule:: skfem.element
+   :members:
+   :exclude-members: DiscreteField, ElementH1, ElementHdiv, ElementHcurl, ElementComposite, ElementVectorH1
