### 2020-11-11 1cef092
Make typing more consistent
