+++ skfem/assembly/basis/interior_basis.py
@@ -32,10 +32,10 @@ class InteriorBasis(Basis):
     def __init__(self,
                  mesh: Mesh,
                  elem: Element,
-                 mapping: Mapping = None,
-                 intorder: int = None,
-                 elements: ndarray = None,
-                 quadrature: Tuple[ndarray, ndarray] = None):
+                 mapping: Optional[Mapping] = None,
+                 intorder: Optional[int] = None,
+                 elements: Optional[ndarray] = None,
+                 quadrature: Optional[Tuple[ndarray, ndarray]] = None):
         """Combine :class:`~skfem.mesh.Mesh` and :class:`~skfem.element.Element`
         into a set of precomputed global basis functions.
 
