### 2020-11-13 61e24bc
nine adaptive refinements, as before, so's not to confuse test
