### 2020-11-23 8c378d9
bump minor version; fixes #518
