### 2020-12-01 e1d1ee7
allow vectorial functionals; fixes #520
