### 2020-12-02 247c0e2
_kernel sums over abscissae, assemble over elements
