### 2020-12-09 0a7899e
vectorize calculation of midpoints
