### 2020-12-09 499d20c
vectorize
