### 2020-12-10 f04513f
need to update MeshLine.facets in _adaptive_refine; fixes #522
